### 2024-08-09 #7096 @editorialbot
[PRE REVIEW]: GASTLI: A Python package for coupled interior–atmosphere modelling of volatile-rich planets
**Submitting author:** @lorenaacuna (<a href="http://orcid.org/0000-0002-9147-7925">Lorena Acuña</a>)
**Repository:** https://github.com/lorenaacuna/GASTLI
**Branch with paper.md** (empty if default branch): main
**Version:** v0.9
