### 2024-08-30 #7096 @JBorrow
Hi @tdsandnes and @jkeger; would either (or both!) of you be interested in reviewing this package for the Journal of Open Source Software? Our reviewers guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html; I'm also happy to answer any questions you have.
