### 2024-08-31 #7096 @tdsandnes
Hi Josh,

I'm in the middle of crunch time for my thesis, so am afraid I'll have to
give this a pass.

Best,
Tom

On Fri, Aug 30, 2024 at 2:52 PM Josh Borrow ***@***.***>
wrote:

> Hi @tdsandnes <https://github.com/tdsandnes> and @jkeger
> <https://github.com/jkeger>; would either (or both!) of you be interested
> in reviewing this package for the Journal of Open Source Software? Our
> reviewers guide is available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html; I'm also
> happy to answer any questions you have.
>
