### 2024-09-06 #7096 @JBorrow
Hi @bobmyhill, @ivalaginja, and @kathlandgren: I found you as a previous corresponding authors for JOSS and think you'd be an excellent fit to review this paper. Would you be interested in that? You can find our reviewers guide here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
