+++ pyproject.toml
@@ -15,6 +15,8 @@ documentation = "https://gastli.readthedocs.io/en/latest/"
 authors = [
     {name = "Lorena Acuna", email = "acuna@mpia.de"},
 ]
+readme = "README.rst"
+license = {file = "LICENSE"}
 dependencies = [
     "numpy >= 1.25.0",
     "pandas >= 2.0.3",
