+++ pyproject.toml
@@ -8,7 +8,7 @@ requires = [
 
 [project]
 name = "gastli"
-version = "0.9.0"
+version = "0.9.1"
 description = "The GAS gianT modeL for Interiors"
 repository = "https://github.com/lorenaacuna/GASTLI"
 documentation = "https://gastli.readthedocs.io/en/latest/"
