### 2024-09-16 1ad17ee
Path to input files are an environment variable
