+++ src/water_curves.py
@@ -3,6 +3,7 @@
 import pandas as pd
 import numpy as np
 from matplotlib import pyplot as plt
+import os
 
 
 
@@ -19,7 +20,8 @@ class water_curves:
 
 
         # Arguments of __init__
-        self.path_to_file = path_to_file
+        #self.path_to_file = path_to_file
+        self.path_to_file = os.environ['GASTLI_input_data_path']
 
         data = pd.read_csv(path_to_file+'Input/Plotting/water_curves.dat', sep='\s+')
 
