### 2024-09-16 a63b05a
Update installation.rst
Update version 0.9.1 is the main branch and stable version.
