### 2024-09-25 #7096 @JBorrow
Hi @MartianColonist, I found you in the JOSS reviewers database. Would you be interested in reviewing this submission? You can find our reviewers guide here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
