### 2024-09-25 #7096 @MartianColonist
Thanks for the invite @JBorrow!
I'm happy to review this JOSS paper. My formal review will happen after the JWST deadline, but I'll endeavour to offer some big picture feedback before that.
