### 2024-09-26 #7288 @editorialbot
[REVIEW]: GASTLI: A Python package for coupled interior–atmosphere modelling of volatile-rich planets
**Submitting author:** @lorenaacuna (<a href="http://orcid.org/0000-0002-9147-7925">Lorena Acuña</a>)
**Repository:** https://github.com/lorenaacuna/GASTLI
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
