### 2024-09-26 #7288 @MartianColonist
@lorenaacuna, I'm looking forward to reviewing your JOSS manuscript and Python package!

One quick thing: the embedded images aren't showing in the notebooks in the docs (probably because they are PDFs). Can you try converting them to pngs to see if they'll render that way?
