### 2024-09-29 lorenaacuna/GASTLI#2 @ivalaginja
Minor requested clarifications in JOSS paper
Hi @lorenaacuna,

as a part of https://github.com/openjournals/joss-reviews/issues/7288, I wanted to drop in here two minor requests for the paper. It is overall very well written and concise, and addresses all the requirements, so both points I bring up here are completely minor>

The first point is that I think it might help if there is some clarification in the part of the paper about the package itself, that describes the parallelization on 8 cores. In the parentheses that follow, it is written:
```
[...] (four at 3.49 GHz and four at 2.42 GHz processors) [...]
```
I wanted to confirm that it is indeed supposed to say "four" for both processor frequency cases, and I also wanted to suggest to clarify four of what exactly you mean.

The second point are the references. All references of Acuña et al. include the author's first name, while none of the other references do. Is this on purpose, and what is the reasoning behind it? From what I know about this citation style, the goal is to be consistent throughout, even for own works of the author. I could be wrong though, so if there is a specific reason for this, please elaborate. Otherwise, I would ask you to change this in the paper and use one consistent reference style.
