### 2024-09-29 #7288 @ivalaginja
@lorenaaacuna, could you elaborate why in the package documentation under https://gastli.readthedocs.io/en/latest/#development, there are six people listed as main developers, while the paper only lists five co-authors? I see that the sixth one is mentioned in the acknowledgments of the paper. How come? This could be totally fine, I am just trying to make sure I understand how the choice was made.
