### 2024-10-17 #7288 @lorenaacuna
@ivalaginja Yes, exactly. Iva Momcheva helped with optimisation and debugging. I invited her to be co-author of this software paper, but she declined and said she preferred to be mentioned in the acknowledgements.
