### 2024-10-17 a23659a
Update tutorial.rst
Changed .pdf files to .png for correct rendering in docs notebooks
