### 2024-10-17 3649c3c
Update retrieval.rst
