### 2024-10-17 #7288 @lorenaacuna
@MartianColonist Thanks so much for catching this. I changed this in the documentation, now you should see the .png version of the plots rendered in the notebooks.
