### 2024-10-17 5f184a3
Update paper.bib
Change citation of Acuna+2024
