+++ docs/JOSS/paper.bib
@@ -554,20 +554,22 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
 @ARTICLE{gastli_science,
-       author = {{Acu{\~n}a}, Lorena and {Kreidberg}, Laura and {Zhai}, Meng and {Molli{\`e}re}, Paul},
-        title = "{GASTLI: An open-source coupled interior-atmosphere model to unveil gas giant composition}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+       author = {{Acu{\~n}a}, L. and {Kreidberg}, L. and {Zhai}, M. and {Molli{\`e}re}, P.},
+        title = "{GASTLI. An open-source coupled interior-atmosphere model to unveil gas-giant composition}",
+      journal = {\aap},
+     keywords = {methods: numerical, planets and satellites: atmospheres, planets and satellites: composition, planets and satellites: interiors, planets and satellites: individual: HAT-P-26 b, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2024,
-        month = jun,
-          eid = {arXiv:2406.10032},
-        pages = {arXiv:2406.10032},
-          doi = {10.48550/arXiv.2406.10032},
+        month = aug,
+       volume = {688},
+          eid = {A60},
+        pages = {A60},
+          doi = {10.1051/0004-6361/202450559},
 archivePrefix = {arXiv},
        eprint = {2406.10032},
  primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240610032A},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...688A..60A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
