### 2024-10-17 8566321
Update paper.md
Clarified number of cores in M2 Apple
