### 2024-10-17 lorenaacuna/GASTLI#2 @lorenaacuna
@ivalaginja thanks for the comments. Yes, it's "four": four cores with 3.49 GHz processors and four cores with 2.42 GHz processors. We have rephrased the parentheses text in the .md doc to be more clear as: "(4 cores operating at 3.49 GHz and the other 4 cores operating at 2.42 GHz)".

For the second point, I think the references that show the first author name are because they were taken from arxiv. I replaced them with the NASA bibtex entry and it was solved.
