### 2024-10-28 lorenaacuna/GASTLI#3 @ivalaginja
Include community guidelines
Please add guidelines to your repository on how to:
1. Contribute to the software
2. Report issues or problems with the software
3. Seek support

All of these should be covered in your documentation. No. 1 would ideally also have a `CONTRIBUTING` file (md or txt) or equivalent on your repository.
