### 2024-10-28 #7288 @ivalaginja
Hey @lorenaacuna, I am slowly working through the checklist. As expected, the installation is the most time-consuming part, but in the meantime, could I ask you to address https://github.com/lorenaacuna/GASTLI/issues/3 please?
