### 2024-10-28 #7288 @ivalaginja
And I opened another issue with the goal to clarify the documentation somewhat: https://github.com/lorenaacuna/GASTLI/issues/4
