### 2024-11-03 lorenaacuna/GASTLI#5 @ivalaginja
Installation clarifications
This is currently the biggest blocker for my review in https://github.com/openjournals/joss-reviews/issues/7288.

## Installation instructions structure

First and foremost, considering that you give very detailed instructions on how to install your package on several OS, and specifically different MacOS processors, I am missing a top-level overview of the installation process for users who don’t need the detailed instructions. For example, this could be:

> The installation of `gastli` requires Python version 9 or higher, with the dependencies `ipython`, `numpy` and `jupyter`, as well as a Fortran compiler. The `gastli` package itself can be installed with pip. You will also need to download a zip file of input data required by the package.

Then I would add a warning/note about the processor details, so that people know up front that this is something to look out for.

Only then would I go into detailed installation instructions, where I think you can separate:
1. The conda environment setup with the installation of Python, its dependencies, and gastli itself - this should be pretty much the same on any OS, provided that a conda distribution is already installed.
2. The input data download, again the same across all OS.
3. A Fortran compiler installation - here is where you need to separate between Linux, MacOS with Intel chips and MacOS with Apple chips.

The result of this would be that users who are experienced enough with similar installations can just get the gist of things and move on. You also warn about the processor bifurcation immediately in the beginning, so users on a MacOS will know to be mindful when starting the installation. And finally, it still gives you the space to provide all the details about the different installation procedures.

## General comments

Apart from the above, below I am providing a couple of review questions on things that are still unclear to me. If I simply didn't understand something well, I apologize - I will be happy to rectify my comments with any new information you can give me.

1. You recommend to install Python, but then you proceed to do a new Python installation in a conda environment for gastli. Why? The system will only ever use one or the other as far as I know, which means that you just want the Python installation in the conda environment I believe?
2. Have you considered providing a conda recipe for the part that requires a conda environment setup? For example something like an `environment.yml` file that would install all the conda installations at once instead of having to type them manually? Since you have very few requirements, you could also just provide the one line (instead of a full `environment.yml` file) to create the conda environment with the right dependencies automatically: `conda create -n gastli python=3.10 ipython numpy jupyter`, and then the line that installs gastli: `pip install gastli==0.9.1`. If you were to use a `environment.yml` file, it would be able to run both of these things at once - the two steps are so simple though that the use of such a file is not necessarily warranted.
3. I also just realized that your `gastli` pip package already lists requirements, so you might also just put any dependencies in there instead of having the user install them one by one. Or have a conda recipe that includes your pip package, as I mentioned in 2). Either way, it would make the installation easier.
4. In the section where you describe the installation with Rosetta, you don’t even mention a dedicated conda environment, you just say to “install the following packages in miniconda3” - package installations should always be directed to a specific environment, in your case the `gastli`/`GASTLI_env` environment.
5. You recommend to work with a conda env but you also recommend to install `gcc` with homebrew instead of conda. Why? There could be a good reason to have a system-wide `gcc` install for your purposes with homebrew, but at first sight, it seems more logical to contain all dependencies for GASTLI in the same conda env. `gcc` seems to be installable through conda-forge, have you tried that?
6. You claim: “The installation of GASTLI on Apple machines with the M1/M2/M3 chip requires Intel emulation with Rosetta.” It would be useful to have a concrete example that triggers the mentioned error so that the user can quickly determine if the Rosetta installation is indeed required for them.
7. Also, does the above mean that option 2) with the disk image doesn’t actually work on M1/M2/M3 machines?
8. I am on a MacOS 12.2 Monterey using an M1 and tried to use the disk image solution because it seems so much easier than to set up Rosetta. However, the disk image installation doesn’t seem to be recognized when I run `import gastli.Coupling as cpl` in an ipython terminal from my gastli conda environment. Did you try an installation like this yourself, is there more to it than just running the installer? For reference, the error message I get:
```
ImportError: dlopen(/Users/ilaginja/opt/miniconda3/envs/gastli3/lib/python3.10/site-packages/gastli/fortinput.cpython-310-darwin.so, 0x0002): Library not loaded: /usr/local/opt/gcc/lib/gcc/current/libgfortran.5.dylib
  Referenced from: /Users/ilaginja/opt/miniconda3/envs/gastli3/lib/python3.10/site-packages/gastli/fortinput.cpython-310-darwin.so
  Reason: tried: '/usr/local/opt/gcc/lib/gcc/current/libgfortran.5.dylib' (no such file), '/usr/local/lib/libgfortran.5.dylib' (no such file), '/usr/lib/libgfortran.5.dylib' (no such file)
```
