### 2024-11-03 #7288 @ivalaginja
The third issue that is part of my review is here: https://github.com/lorenaacuna/GASTLI/issues/5
It addresses the installation instructions.
