### 2024-11-04 5b4769e
Create CONTRIBUTING.md
