+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contributing to GASTLI
+
+Thank you for your interest in contributing to GASTLI! Below are guidelines on how you can report issues and contribute code.
+
+## Reporting issues
+To report bugs or suggest new features, please create a [new issue](https://github.com/lorenaacuna/GASTLI/issues) in Github. 
+When reporting an issue:
+
+- Provide details about your operating system, GASTLI version and Python version.
+- Whenever possible, please also include a brief, self-contained code example that demonstrates the problem. 
+
+**Reporting the issue on Github is the preferred method** as it makes the issue public, helping others who might encounter the same problem. However, if you have questions about modelling or need specific guidance, you can also contact the main developer at [acuna@mpia.de](mailto:acuna@mpia.de).
+
+## Contributing code 
+We welcome code contributions to GASTLI. To get started:
+- **Gauge feasibility**: Before starting, please reach out to the main developer to gauge whether your proposed feature aligns with GASTLI's applications.
+- **Implementation guidance**: Once it is confirmed that your proposed feature is feasible, the main developer can help guide the process, making it easier to integrate your work into GASTLI.
