### 2024-11-04 7849196
Update index.rst
Added "contributing" section to documentation
