+++ docs/source/index.rst
@@ -39,6 +39,7 @@ Contents
    installation
    atm_grid
    test_suite
+   ../../CONTRIBUTING.md
 
 .. toctree::
    :maxdepth: 1
