### 2024-11-04 d06d837
Create contributing.rst
