### 2024-11-04 lorenaacuna/GASTLI#3 @lorenaacuna
Thanks for this suggestion @ivalaginja. I added a CONTRIBUTING.md file and a "Contributing" section in the user guide in the documentation. I close the issue.
