### 2024-11-04 lorenaacuna/GASTLI#4 @lorenaacuna
Thanks for the suggestions @ivalaginja. To address point 1), I grouped all the sub-sections (i.e "Installation", "Test suite", "Contributing", "Quickstart", etc) under a broader section titled "User Guide". Consequently, I removed the previous distinction between "User Guide" and "Tutorials". All of these sub-sections now have links in the sidebar and on the welcome page. I hope it is more clear now.
