+++ docs/source/tutorial.rst
@@ -5,6 +5,12 @@ Quickstart
 Welcome to GASTLI's quickstart tutorial. In this tutorial we explain how to compute interior structure and atmospheric models
 and plot the results. These include the internal and atmospheric profiles, a pressure-temperature diagram, mass-radius relations, and thermal evolution curves. In the last section of this tutorial, we also discuss how to check and ease convergence. Here's an overview of this page:
 
+.. toctree::
+   :ref: `_intatmobject<Creating an interior-atmosphere object>`.
+   Generating a thermal evolution curve
+   Convergence
+
+
 .. _intatmobject:
 
 Creating an interior-atmosphere object
