+++ docs/source/thermal_evol.rst
@@ -1,0 +1,154 @@
+
+
+Generating a thermal evolution curve
+=========================================
+
+To obtain the internal temperature (or luminosity) and radius as a function of age, we need to use the GASTLI class ``Thermal_evolution``. This class obtains a sequence of static interior-atmosphere models at different internal temperatures with the function ``thermal_evolution_class_object.main()``. The input array ``Tint_array`` specifies the discreet internal temperatures at which the static models are computed. We recommend to save the sequence of interior models in a file, as in the example below. In this example, we name the thermal evolution class object ``my_therm_obj``. The output of the ``main()`` thermal class function is:
+
+- The derivative of the entropy **dS/dt** in SI units: ``thermal_evolution_class_object.f_S``
+- The envelope **entropy at 1000 bar** in J/kg/K: ``thermal_evolution_class_object.s_top_TE``
+- The envelope mean entropy in J/kg/K: ``thermal_evolution_class_object.s_mean_TE``
+- The **total planet radius**, and interior radius (center to surface pressure) in Jupiter radii: ``thermal_evolution_class_object.Rtot_TE`` and ``thermal_evolution_class_object.Rbulk_TE``
+- The surface temperature in K: ``thermal_evolution_class_object.Tsurf_TE``
+
+
+.. code-block:: python
+
+   # Import GASTLI thermal module
+   import gastli.Thermal_evolution as therm
+   # Other Python modules
+   import numpy as np
+   # Path to input files
+   # Dont forget the "/" at the end of the string
+   path_input = "/Users/acuna/Desktop/gastli_input_data/"
+   # Create thermal evolution class object
+   my_therm_obj = therm.thermal_evolution(path_to_file=path_input)
+   # Input for interior
+   M_P = 18.76     # Earth units
+   # Equilibrium temperatures
+   Teqpl = 1000.
+   # Core mass fraction
+   CMF = 0.5
+   log_FeH = np.log10(20.) # 20 x solar
+   Tint_array = np.asarray([50.,60.,70.,80., 100., 110., 120., 130., 140., 160., 150., 160., 200., 240., 300.])
+   # Run sequence of interior models at different internal temperatures
+   my_therm_obj.main(M_P, CMF, Teqpl, Tint_array, log_FeH=log_FeH)
+   # Recommended: save sequence of interior models in case thermal evol eq. solver stops
+   Rjup = 11.2  # Jupiter radius in Earth units
+   data = np.zeros((len(my_therm_obj.f_S),7))
+   data[:,0] = my_therm_obj.f_S
+   data[:,1] = my_therm_obj.s_mean_TE
+   data[:,2] = my_therm_obj.s_top_TE
+   data[:,3] = my_therm_obj.Tint_array
+   data[:,4] = my_therm_obj.Rtot_TE*Rjup
+   data[:,5] = my_therm_obj.Rbulk_TE*Rjup
+   data[:,6] = my_therm_obj.Tsurf_TE*Rjup
+   fmt = '%1.4e','%1.4e','%1.4e','%1.4e','%1.4e','%1.4e','%1.4e'
+   np.savetxt('thermal_sequence_HATP26b_CMF50_20xsolar.dat', data,header='f_S s_mean_TE s_top_TE Tint Rtot Rbulk    Tsurf',comments='',fmt=fmt)
+
[…]
