+++ docs/source/int_atm_object.rst
@@ -1,0 +1,451 @@
+
+
+Creating an interior-atmosphere object
+======================================
+
+You can start a simple interior-atmosphere object to generate one model by initialising the ``coupling`` class:
+
+.. code-block:: python
+
+   # Import coupling module
+   import gastli.Coupling as cpl
+   # Other Python modules
+   import numpy as np
+   # Specify path to input data folder
+   path_to_input = "/Users/acuna/Desktop/gastli_input_data/"
+   # Initialise coupling class
+   my_coupling = cpl.coupling(path_to_file=path_to_input)
+
+
+.. important:: Replace the variable ``path_to_input`` in the code above with the absolute path to the input data folder in your computer. Do not forget the two slash symbols (``/``), one at the beginning and the other one at end of the string.
+
+Next, specify the input variables for the interior-atmosphere model:
+
+.. code-block:: python
+
+   # Input for interior
+   # 1 Mjup in Mearth units
+   M_P = 318.
+   # Internal temperature
+   Tintpl = 99.
+   # Equilibrium temperature
+   Teqpl = 122.
+   # Core mass fraction
+   CMF = 0.01
+
+If the log-metallicity of the atmosphere, :math:`log(Fe/H)`, is known, you must specify it as:
+
+.. code-block:: python
+
+   # Envelope log-metallicity is solar
+   log_FeHpl = 0
+   # C/O ratio is solar
+   CO_planet = 0.55
+   # Run coupled interior-atmosphere model
+   my_coupling.main(M_P, CMF, Teqpl, Tintpl, CO=CO_planet, log_FeH=log_FeHpl)
+
+In a nutshell, you can print the output radius, total metal content, surface gravity, etc:
+
+.. code-block:: python
+
+   print("Case 1, log(Fe/H) is known")
+   # Composition input
+   print("log(Fe/H) atm [x solar] (input) = ",my_coupling.myatmmodel.log_FeH)
+   print("C/O atm (input) = ",my_coupling.myatmmodel.CO)
+   # Output
+   print("Zenv (output) = ",my_coupling.myatmmodel.Zenv_pl)
+   print("Total planet mass M [M_earth] = ",my_coupling.Mtot)
+   print("Temperature at 1000 bar [K] = ",my_coupling.T_surf)
+   print("Planet bulk radius [R_jup] = ",my_coupling.Rbulk_Rjup)
+   print("log10_g: Planet surface gravity (1000 bar) [cm/s2] = ",np.log10(my_coupling.g_surf_planet))
+   print("Total planet radius [R_jup] = ",my_coupling.Rtot)
+   tmm = my_coupling.Mtot*CMF + my_coupling.Mtot*(1-CMF)*my_coupling.myatmmodel.Zenv_pl
+   print("Total metal mass [M_earth] = ",tmm)
+
+   
+You will obtain the following output:
+
+.. code-block:: language
+
+   Case 1, log(Fe/H) is known
+   log(Fe/H) atm [x solar] (input) =  0
+   C/O atm (input) =  0.55
+   Zenv (output) =  0.013532983488449907
+   Total planet mass M [M_earth] =  318.0389824070297
+   Temperature at 1000 bar [K] =  1321.0792698333128
+   Planet bulk radius [R_jup] =  0.9732321589894197
+   log10_g: Planet surface gravity (1000 bar) [cm/s2] =  3.417144909504209
+   Total planet radius [R_jup] =  0.9795696861509946
[…]
