+++ docs/source/convergence.rst
@@ -1,0 +1,106 @@
+
+Convergence
+===========
+
+There are a few input parameters that can help ease convergence of the interior-atmosphere coupling algorithm. These are:
+
+- The **radius initial guess** in Earth radii: ``Rguess`` (default value is 11.2 Earth radii)
+- The **mass power law**: ``pow_law_formass`` (default value is 0.32)
+- The **tolerance**: ``tolerance`` (default value is 0.001)
+
+``Rguess`` and ``tolerance`` are specified in the ``main()`` functions of the coupling and thermal evolution classes, while ``pow_law_formass`` is defined when the class is created. If the algorithm does not converge after 20 iterations, the power law will be reduced by 0.05 until convergence is reached. In that case, a message similar to the following will be displayed: 
+
+.. code-block:: language
+
+   Warning in Coupling.py: The number of interior-atmosphere iterations is greater than 20 
+   The current relative difference between radii is 0.00294796
+   Readjusting mass power law to 0.315
+
+We recommend increasing the tolerance only as a last resort, as this may risk producing a least accurate interior-atmosphere model.
+
+On another note, a high ``pow_law_formass`` can lead to the following error, which concerns only the interior model (not the interior-atmosphere algorithm):
+
+.. code-block:: language
+
+   Error in interior structure model (Fortran): Mass of core layer is NaN
+   This is likely due to a spatial resolution of the radius grid that is
+   too small to resolve the size of the core
+   Increase the resolution of the grid by setting a lower value for the
+   input parameter pow_law_formass
+
+After showing this message, the program execution will stop. We recommend decreasing the power law parameter by 0.05 manually. This may be needed for planets with low masses (i.e Neptune or sub-Neptune mass) and/or with high metal content. The ``pow_law_formass`` parameter should not be lower than 0.29.
+
+Finally, there is a parameter that controls the maximum number of the interior model (not the interior-atmosphere algorithm), ``j_max``. This is specified in the ``main()`` functions of the coupling and thermal evolution classes. The default value is 30, which works for most models with surface pressure ``P_surf`` = 1000 bar. For lower surface pressures, or planets with very low surface gravity (see example below), the interior model may need more than 30 iterations to reach convergence. In that case, the following message will be displayed: 
+
+.. code-block:: language
+
+   Running interior structure model
+   [i] Allowed maximum number of iterations reached.
+
+
+We recommend increasing ``j_max``, to no more than ``j_max = 99``. Sometimes, even if ``j_max`` is in its maximum value and the final radius has converged, the message above can be displayed. This is due to the fact that the interior model uses a precision of 1e-5 to check if the interior structure profiles have converged. A good way to check if this is the case is to plot the evolution of the planet radius and layer interfaces with iteration number. This information is in the ``myplanet.intrf_hist`` and ``myplanet.iter_num`` arrays.
+
+.. code-block:: python
+
+   # Import GASTLI modules
+   import gastli.Coupling as cpl
+   import gastli.constants as cte
+   # Other Python modules
+   import numpy as np
+   import matplotlib.pyplot as plt
+   # Create coupling class
[…]
