+++ docs/source/tutorial.rst
@@ -6,12 +6,10 @@ Welcome to GASTLI's quickstart tutorial. In this tutorial we explain how to comp
 and plot the results. These include the internal and atmospheric profiles, a pressure-temperature diagram, mass-radius relations, and thermal evolution curves. In the last section of this tutorial, we also discuss how to check and ease convergence. Here's an overview of this page:
 
 .. toctree::
-   :ref: `_intatmobject<Creating an interior-atmosphere object>`.
-   Generating a thermal evolution curve
-   Convergence
+   int_atm_object
+   thermal_evol
+   convergence
 
-
-.. _intatmobject:
 
 Creating an interior-atmosphere object
 ======================================
