+++ docs/source/tutorial.rst
@@ -3,717 +3,11 @@ Quickstart
 ==========
 
 Welcome to GASTLI's quickstart tutorial. In this tutorial we explain how to compute interior structure and atmospheric models
-and plot the results. These include the internal and atmospheric profiles, a pressure-temperature diagram, mass-radius relations, and thermal evolution curves. In the last section of this tutorial, we also discuss how to check and ease convergence. Here's an overview of this page:
+and plot the results. These include the internal and atmospheric profiles, a pressure-temperature diagram, mass-radius relations, and thermal evolution curves. In the last section of this tutorial, we also discuss how to check and ease convergence. 
+
+Here's an overview of this quickstart tutorial:
 
 .. toctree::
    int_atm_object
    thermal_evol
    convergence
-
-
-Creating an interior-atmosphere object
-======================================
-
-You can start a simple interior-atmosphere object to generate one model by initialising the ``coupling`` class:
-
-.. code-block:: python
-
-   # Import coupling module
-   import gastli.Coupling as cpl
-   # Other Python modules
-   import numpy as np
-   # Specify path to input data folder
-   path_to_input = "/Users/acuna/Desktop/gastli_input_data/"
-   # Initialise coupling class
-   my_coupling = cpl.coupling(path_to_file=path_to_input)
-
-
-.. important:: Replace the variable ``path_to_input`` in the code above with the absolute path to the input data folder in your computer. Do not forget the two slash symbols (``/``), one at the beginning and the other one at end of the string.
-
-Next, specify the input variables for the interior-atmosphere model:
-
-.. code-block:: python
-
-   # Input for interior
-   # 1 Mjup in Mearth units
-   M_P = 318.
-   # Internal temperature
-   Tintpl = 99.
-   # Equilibrium temperature
-   Teqpl = 122.
-   # Core mass fraction
-   CMF = 0.01
-
-If the log-metallicity of the atmosphere, :math:`log(Fe/H)`, is known, you must specify it as:
-
-.. code-block:: python
-
-   # Envelope log-metallicity is solar
-   log_FeHpl = 0
-   # C/O ratio is solar
-   CO_planet = 0.55
-   # Run coupled interior-atmosphere model
-   my_coupling.main(M_P, CMF, Teqpl, Tintpl, CO=CO_planet, log_FeH=log_FeHpl)
-
-In a nutshell, you can print the output radius, total metal content, surface gravity, etc:
-
-.. code-block:: python
-
-   print("Case 1, log(Fe/H) is known")
-   # Composition input
-   print("log(Fe/H) atm [x solar] (input) = ",my_coupling.myatmmodel.log_FeH)
-   print("C/O atm (input) = ",my_coupling.myatmmodel.CO)
-   # Output
-   print("Zenv (output) = ",my_coupling.myatmmodel.Zenv_pl)
-   print("Total planet mass M [M_earth] = ",my_coupling.Mtot)
-   print("Temperature at 1000 bar [K] = ",my_coupling.T_surf)
-   print("Planet bulk radius [R_jup] = ",my_coupling.Rbulk_Rjup)
[…]
