### 2024-11-04 lorenaacuna/GASTLI#4 @lorenaacuna
Comment 2): I added a table of contents to the Quickstart section in the documentation
