+++ src/Coupling.py
@@ -8,8 +8,9 @@ import math
 import sys
 
 class coupling:
-    def __init__(self,path_to_file,name_grid=None,pow_law_formass=0.32,j_max=30):
-        r"""Class defining objects to run one interior-atmosphere coupled model
+        r"""Class defining objects to run one interior-atmosphere coupled model. 
+            When this class is defined, it automatically initialises interior and atmosphere models and
+            defines required constants.
 
         Args:
             path_to_file:
@@ -20,11 +21,8 @@ class coupling:
                 core mass fraction is very low (< 0.03 approx.) and/or planet is low mass (15-20 Earth masses approx.)
         """
 
-        '''
-        - Initialises interior and atm. models
-        - Other constants
-        '''
-
+    
+    def __init__(self,path_to_file,name_grid=None,pow_law_formass=0.32,j_max=30):
         self.pow_law_formass = pow_law_formass
         self.path_to_file = path_to_file
         self.j_max = j_max
