### 2024-11-04 a741e33
Update GASTLI.py
Updated docstrings in classes GASTLI int and Coupling to show properly in API
