### 2024-11-04 29c0efb
Update Coupling.py
