+++ src/Coupling.py
@@ -8,7 +8,7 @@ import math
 import sys
 
 class coupling:
-        r"""Class defining objects to run one interior-atmosphere coupled model. 
+    r"""Class defining objects to run one interior-atmosphere coupled model. 
             When this class is defined, it automatically initialises interior and atmosphere models and
             defines required constants.
 
@@ -19,8 +19,8 @@ class coupling:
                 power exponent to estimate the initial guess of the planet radius in the interior model.
                 Default is 0.32. Increase if planet is very massive (greater than 5 Jupiter masses aprox). Decrease if
                 core mass fraction is very low (< 0.03 approx.) and/or planet is low mass (15-20 Earth masses approx.)
-        """
-    
+    """
+
     def __init__(self,path_to_file,name_grid=None,pow_law_formass=0.32,j_max=30):
         self.pow_law_formass = pow_law_formass
         self.path_to_file = path_to_file
