### 2024-11-04 lorenaacuna/GASTLI#4 @lorenaacuna
Comment 3): There was a doctoring for this class, but it was placed after the `def __init__(...)` line, so it wasn't showing. Now it is corrected and it shows in the API's automatically generated documentation.
