### 2024-11-05 95aee68
Update index.rst
Extended statement of need in welcome page
