+++ docs/source/index.rst
@@ -12,7 +12,7 @@ Key features:
 - **Up-to-date equations of state (EOS) tables**: GASTLI integrates the latest non-ideal EOS for H/He, water and rock to calculate the density and Grüneisen parameter (adiabatic gradient), which is essential for accurate modelling of high-pressure planetary envelopes. Refer to `this paper <https://www.aanda.org/articles/aa/full_html/2024/08/aa50559-24/aa50559-24.html>`_ for available EOS. If you wish to use other EOS in your models, please check our :doc:`contributing` section.
 - **Flexible mass, composition and irradiation input**: GASTLI supports a wide range of core mass fractions, envelope metal content (i.e. atmospheric metallicity), masses, and irradiation conditions - provided the atmospheric grid covers relevant surface gravities, atmospheric metallicity and equilibrium temperatures. The default atmospheric grid's applicability is covered in our :doc:`atm_grid` section.
 - **Customizable atmospheric grid**: The default atmospheric grid corresponds to clear, 1D atmospheres in chemical equilibrium. If you wish to change these assumptions for your models (e.g., clouds, 3D effects, chemical disequilibrium) or our default atmospheric grid does not meet your required surface gravities, compositions and equilibrium temperatures, GASTLI allows you to input a custom atmospheric grid. For details on grid file format and requirements, check our :doc:`atm_grid` section.
-- **Retrievals**: GASTLI is optimized to compute numerous models within a grid, allowing interpolation for retrievals on mass, radius, atmospheric metallicity, age, and/or internal temperature. See our :doc:`retrievals` section for a comprehensive guide on performing retrievals with GASTLI.
+- **Retrievals**: GASTLI is optimized to compute numerous models within a grid, allowing interpolation for retrievals on mass, radius, atmospheric metallicity, age, and/or internal temperature. See our :doc:`retrieval` section for a comprehensive guide on performing retrievals with GASTLI.
 
 
 License and how to cite
