### 2024-11-05 lorenaacuna/GASTLI#4 @lorenaacuna
Comment 4): extended the text in the welcome page with target audience and a list of key features.
