### 2024-11-05 lorenaacuna/GASTLI#5 @lorenaacuna
Thanks @ivalaginja for the feedback. I start by addressing comment 8):

> I am on a MacOS 12.2 Monterey using an M1 and tried to use the disk image solution because it seems so much easier than to set up Rosetta. However, the disk image installation doesn’t seem to be recognized when I run import gastli.Coupling as cpl in an ipython terminal from my gastli conda environment. Did you try an installation like this yourself, is there more to it than just running the installer? For reference, the error message I get:

This is the message we get if we try to install GASTLI in Apple Silicon with the gfortran native ARM64 installation. This is not explicitly said in the disk image subsection of the installation documentation, but the version I think you need to install of gfortran is the Intel one, not the ARM64. I need to confirm this with the co-author that wrote this sub-section of the documentation. In the meantime, can you try to uninstall your current ARM64 gfortran and install the Intel one? If that does not work, please let me know.

I continue to address the other questions in the following comments.
