### 2024-11-05 37e3a4d
Clarified and re-structured installation doc page
