+++ docs/source/installation.rst
@@ -11,225 +11,54 @@ To install GASTLI, you need to:
 - have a fortran compiler, for example ``gfortran``,
 - download the zip file for the input data `here <https://www.dropbox.com/scl/fi/p2kawqp8gtzh5psn21tjc/gastli_input_data.zip?rlkey=fc0mfxvpck5mukkqhk1f8hkad&st=ggsa4zmk&dl=0>`_, and place it in a directory in your computer that is easy to find
 
-If you do not have the first two requirements, we offer an installation guide below.
+If you do not have either of the first two requirements, we offer an installation guide for different OS below:
+
+.. toctree::
+   linux_install
+   macos_intel_install
+   macos_arm64_install
 
 
-.. important::
+.. warning::
 
-   The latest stable version is 0.9.1, which corresponds to the main branch. Version 0.9.2 is the development branch. We recommend users to make sure that the install version 0.9.1 if they use pip ``pip install gastli==0.9.1``
-
-
-Fortran and Python installation
--------------------------------
-
-Linux
-~~~~~
-
-On Linux, you can install Python and the fortran compiler with:
-
-.. code-block:: bash
-
-    sudo apt-get install python python-pip gfortran
-
-Depending on the distribution, ``python`` may need to be replaced with ``python3``.
-
-Mac OS (all)
-~~~~~~~~~~~~
-
-For Mac OS, we recommend to use `homebrew <https://brew.sh>`_. Trying to install gfortran (or gcc) from source can be complicated, and setup-related errors can be easily introduced. 
-
-For a safe installation, start executing in your terminal:
-
-.. code-block:: bash
-
-    brew update
-    brew upgrade
-    brew doctor
-
-The command ``brew doctor`` may show a list of fixes. We recommend to go through this list before moving on to the next step.
-
-Finally, the fortran compiler can be installed by running in the terminal: 
-
-.. code-block:: bash
-
-    brew install gcc
-
-Mac OS (Apple Silicon M1/M2/M3)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-If your Mac OS has an Apple Silicon chip, and you follow the steps above for Mac OS, you may encounter the following error when trying to access any variable, class or function of the package:
-
-.. code-block:: bash
+   Mac OS with Apple Silicon processors (M1, M2 and M3) have ARM64 as the native architecture. These systems have shown problems when installing 
+   a Python package that compiles a fortran backend because Python's default fortran compiler uses Intel architecture. One way to recognise this is
+   to try installing GASTLI's with pip (or from source) anfd testing the installation as indicated below. If you get a message such as the following, 
+   you need to follow the instructions for Apple Silicon in the :doc:`macos_arm64_install` section:
+   
+   .. code-block:: python
 
     Traceback (most recent call last):
     File "<stdin>", line 1, in <module>
     ImportError: dlopen(/Users/acuna/anaconda3/envs/GASTLI_norosetta/lib/python3.10/site-packages/gastli/dimensions.cpython-310-darwin.so, 0x0002): symbol not found in flat namespace (_f2pyinitdimensions_)
 
-There are two options to fix this error: 1) using an Intel emulation with Rosetta, or 2) using a disk image. We explain both options below.
-
-**1) Installation with Rosetta**
-
-The installation of GASTLI on Apple machines with the M1/M2/M3 chip requires Intel emulation with Rosetta.
-
-.. code-block:: bash
-
-   softwareupdate --install-rosetta
-
[…]
