+++ docs/source/macos_arm64_install.rst
@@ -1,0 +1,153 @@
+Fortran installation in Mac OS (Apple Silicon M1/M2/M3)
+-----------------------------------
+
+If you are reading this section, it is because the you have an Apple Silicon chip and encountered the following error when trying to access any variable, class or function of the package:
+
+.. code-block:: bash
+
+    Traceback (most recent call last):
+    File "<stdin>", line 1, in <module>
+    ImportError: dlopen(/Users/acuna/anaconda3/envs/GASTLI_norosetta/lib/python3.10/site-packages/gastli/dimensions.cpython-310-darwin.so, 0x0002): symbol not found in flat namespace (_f2pyinitdimensions_)
+
+There are two options to fix this error: 1) using an Intel emulation with Rosetta, or 2) using a disk image. We explain both options below.
+
+**1) Installation with Rosetta**
+
+The installation of GASTLI on Apple machines with the M1/M2/M3 chip requires Intel emulation with Rosetta.
+
+.. code-block:: bash
+
+   softwareupdate --install-rosetta
+
+Now go to the Applications folder and find the iTerm icon. Make a copy of this application and name the new copy as something like "iTerm_Rosetta". Right click iTerm_Rosetta, choose "Get Info", and select the "Open using Rosetta" box. To test that you are indeed using the Intel emulator, type the following in your iTerm_Rosetta:
+
+.. code-block:: bash
+
+   arch
+
+This command should return ``i386``.
+
+Next, install homebrew with Rosetta:
+
+.. code-block:: bash
+
+   /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+
+With the Intel emulation, Homebrew should be installed at ``/usr/local/bin/brew``. Add the following to your ``.bash_profile``
+
+.. code-block:: bash
+
+   alias brew_i386="/usr/local/bin/brew"
+
+In the future, you will use `brew_i386` as an alternative of `brew` with the Intel emulation.
+
+For completeness only, you might also install Homebrew in your M1/M2/M3 terminal, which should be then installed at ``/opt/homebrew/bin/brew``. Add the following to your ``.bash_profile``
+
+.. code-block:: bash
+
+   alias brew="/opt/homebrew/bin/brew"
+
+Now we will install ``miniconda3`` in Rosetta, but before that, we will have to modify ``.bash_profile`` so we could handle the ``conda`` between M1/M2/M3 and Rosetta separately. Here I assume you already installed ``anaconda`` in your M1/M2/M3 terminal, so the following block should be in your ``.bash_profile``:
+
+.. code-block:: bash
+
+   # >>> conda initialize >>>
+   # !! Contents within this block are managed by 'conda init' !!
+   __conda_setup="$('/Users/xxxx/anaconda3/bin/conda' 'shell.bash' 'hook' 2> /dev/null)"
+   if [ $? -eq 0 ]; then
+      eval "$__conda_setup"
+   else
+      if [ -f "/Users/xxxx/anaconda3/etc/profile.d/conda.sh" ]; then
+          . "/Users/xxxx/anaconda3/etc/profile.d/conda.sh"
+      else
+          export PATH="/Users/xxxx/anaconda3/bin:$PATH"
+      fi
+  fi
+  unset __conda_setup
+  # <<< conda initialize <<<
+
+Note that the "xxxx" here should be your username. Let's cut these few lines and paste them into a separate file ``.init_conda_arm64.sh`` in the home directory. We will come back to handle this file later.
+
[…]
