+++ docs/source/linux_install.rst
@@ -1,0 +1,12 @@
+
+Fortran installation in Linux
+-------------------------------
+
+On Linux, you can install Python and the fortran compiler with:
+
+.. code-block:: bash
+
+    sudo apt-get install python python-pip gfortran
+
+Depending on the distribution, ``python`` may need to be replaced with ``python3``.
+
