+++ docs/source/installation.rst
@@ -9,7 +9,7 @@ To install GASTLI, you need to:
 
 - have a Python 3.9+ installation,
 - have a fortran compiler, for example ``gfortran``,
-- **download the zip file for the input data `here <https://www.dropbox.com/scl/fi/p2kawqp8gtzh5psn21tjc/gastli_input_data.zip?rlkey=fc0mfxvpck5mukkqhk1f8hkad&st=ggsa4zmk&dl=0>`_,** and place it in a directory in your computer that is easy to find
+- **download the zip file for the input data** `here <https://www.dropbox.com/scl/fi/p2kawqp8gtzh5psn21tjc/gastli_input_data.zip?rlkey=fc0mfxvpck5mukkqhk1f8hkad&st=ggsa4zmk&dl=0>`_, and place it in a directory in your computer that is easy to find
 
 If you do not have either of the first two requirements, we offer an installation guide for different OS below:
 
