### 2024-11-05 00e5b3f
Update macos_arm64_install.rst
