### 2024-11-05 b627114
Update macos_arm64_install.rst
