### 2024-11-05 2019cf1
Update macos_arm64_install.rst
