### 2024-11-05 lorenaacuna/GASTLI#5 @lorenaacuna
Answer to comments 1), 2) and 3):

> You recommend to install Python, but then you proceed to do a new Python installation in a conda environment for gastli. Why? The system will only ever use one or the other as far as I know, which means that you just want the Python installation in the conda environment I believe?

Exactly. We recommend the user to use conda or venv environments to avoid conflicts, but it is up to the user to decide this. I have modified that part of the text as "If you decide to create a new conda environment (for example, “GASTLI_env”), [...]" just to clarify that the use of conda is optional.

> Since you have very few requirements, you could also just provide the one line (instead of a full environment.yml file) to create the conda environment with the right dependencies automatically: conda create -n gastli python=3.10 ipython numpy jupyter, and then the line that installs gastli: pip install gastli==0.9.1.

ipython and jupyter are not required to use GASTLI, I removed these from the docs. Now the user just needs to 1) create a conda/venv environment (if they choose to) and 2) run pip install gastli==0.9.1

> I also just realized that your gastli pip package already lists requirements, so you might also just put any dependencies in there instead of having the user install them one by one.

Yes, this is why the .yml file is not so necessary. numpy, which was mentioned in the documentation before, is already in the requirements. I removed the mention of numpy together with ipython and jupyter.
