### 2024-11-05 de99c5c
Update macos_arm64_install.rst
