### 2024-11-05 lorenaacuna/GASTLI#5 @lorenaacuna
Answers to comments 6) and 7):

> You claim: “The installation of GASTLI on Apple machines with the M1/M2/M3 chip requires Intel emulation with Rosetta.” It would be useful to have a concrete example that triggers the mentioned error so that the user can quickly determine if the Rosetta installation is indeed required for them.

I added instructions on how to check if the user needs to follow the instructions for Apple Silicon in the very first warning in the installation front page. This includes an example of the error message that we get when trying to install GASTLI in the arm64 architecture.

> Also, does the above mean that option 2) with the disk image doesn’t actually work on M1/M2/M3 machines?

No, the disk image solution is an alternative to Rosetta for Apple Silicon users, so it should work on the M1-M3 machines.
