### 2024-11-05 lorenaacuna/GASTLI#5 @lorenaacuna
@ivalaginja I will close this issue once you confirm that you could install it in your machine. I also want to wait for the co-author that wrote the Apple Silicon installation with disk images to clarify this part of the documentation. He'll get back next week.
