### 2024-11-05 #7288 @lorenaacuna
@JBorrow I'm waiting for a co-author to get back to us to clarify one last comment in [issue #5](https://github.com/lorenaacuna/GASTLI/issues/5) before I close it.

I'm looking forward to @MartianColonist's review.
