### 2024-11-06 lorenaacuna/GASTLI#5 @ivalaginja
@lorenaacuna thanks a lot for the detailed answers! I will look at them in detail shorty, it looks like all of my questions were addressed. I'll let you know!
