### 2024-11-06 lorenaacuna/GASTLI#3 @ivalaginja
@lorenaacuna thanks for adding this! I won't block the JOSS review for this since you have added the guidelines now, but just as a general suggestion I would recommend to be as detailed in the "Contributing code" part as you are in the "Reporting issues" part - something like (1) Fork the repo,  (2) Create a branch off `main`, (3) Implement your changes, (4) ...

It just makes it easier, but it is not strictly necessary. Closed topic for me! 👍
