### 2024-11-06 lorenaacuna/GASTLI#4 @ivalaginja
I like it a lot better, thank you for addressing this!
