### 2024-11-06 lorenaacuna/GASTLI#5 @ivalaginja
> I changed the installation page in the documentation to address the suggestions for the structure. The installation page now follows the structure:
>
>     * overview of requirements
>
>     * table of content for Fortran installation for different OS
>
>     * warning for Apple Silicon Users
>
>     * usual installation with conda and pip. We recommend the users to use these.

This is so much better now, amazing! :D
