### 2024-11-06 lorenaacuna/GASTLI#5 @ivalaginja
> Answer to comment 4):
>
> > In the section where you describe the installation with Rosetta, you don’t even mention a dedicated conda environment, you just say to “install the following packages in miniconda3” - package installations should always be directed to a specific environment, in your case the gastli/GASTLI_env environment.
>
> First the steps to install miniconda3 need to be followed. Then once the user has miniconda3 working with Rosetta, they can install GASTLI as usual. I removed the package part and added a clarification at the end that the user can create a conda/venv environment to avoid conflicts.

This is also a lot clearer in the docs now, thank you!
