### 2024-11-06 lorenaacuna/GASTLI#5 @ivalaginja
> Answer to comment 5):
>
> > You recommend to work with a conda env but you also recommend to install gcc with homebrew instead of conda. Why? There could be a good reason to have a system-wide gcc install for your purposes with homebrew, but at first sight, it seems more logical to contain all dependencies for GASTLI in the same conda env. gcc seems to be installable through conda-forge, have you tried that?
>
> The installation of gcc/gfortran is quite tricky. If the user chooses to install gcc through conda-forge instead of homebrew, the compilation of the fortran backend done by Python's packages (that is meson and pip in the case of GASTLI) may not work, depending on the OS. We noticed that using homebrew is more successful in these cases, so this is what we recommend.

Aha! Good to know. Thanks for explaining!
