### 2024-11-24 lorenaacuna/GASTLI#5 @ivalaginja
I am currently going through the installation with Rosetta  for a M1/M2/M3 Mac and have a couple of questions and comments:

1. Twice in the installation instructions, you instruct to "cut and paste" -> do you really mean cut, or do you mean copy and paste? If it's indeed cut, it might be worth to clarify in parentheses `(copy and remove)`, if you don't actually mean cut, then replace the word with paste. I have to admit that as someone who has never used Rosetta before, it is unclear which of the two is needed.

The two lines are:
- https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L70
- https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L103

2. [Here](https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L51), you say you'll assume people already have anaconda installed in their main terminal. Since most people are very confused between anaconda, miniconda and conda, here again it would be good to specify something like: "[...] you already installed ``anaconda`` or ``miniconda``  in your M1/M2/M3 terminal [...]".

3. Claim: https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L129 - but how will I actually know? Might be worth adding half a sentence describing how exactly the user will know.

4. The most important comment is, actually, that when I ran `softwareupdate --install-rosetta`, I get the following output - note the error message in the second to last line before it claims it finished successfully:
```bash
maintenncesMBP6:~ ilaginja$ softwareupdate --install-rosetta
I have read and agree to the terms of the software license agreement. A list of Apple SLAs may be found here: http://www.apple.com/legal/sla/
Type A and press return to agree: A
2024-11-24 20:33:44.639 softwareupdate[39894:12294479] Package Authoring Error: 002-57052: Package reference com.apple.pkg.RosettaUpdateAuto is missing installKBytes attribute
Install of Rosetta 2 finished successfully
```
The result is that I do not have the iTerm app on my machine. I tried googling this but wasn't able to understand what is happening. Would you have any clues?
