### 2024-11-25 lorenaacuna/GASTLI#5 @lorenaacuna
> This is the message we get if we try to install GASTLI in Apple Silicon with the gfortran native ARM64 installation. This is not explicitly said in the disk image subsection of the installation documentation, but the version I think you need to install of gfortran is the Intel one, not the ARM64. I need to confirm this with the co-author that wrote this sub-section of the documentation. In the meantime, can you try to uninstall your current ARM64 gfortran and install the Intel one? If that does not work, please let me know.

My co-author got back about the disk image solution: If you have the M1/M2/M3 Mac OS, you need to install the ARM64 (not Intel, as I said in a previous message). For them, the disk image worked just "out of the box", following the instructions in FX Coudert's repository.

@ivalaginja we suggest to check what fortran binary is being called by default by running the command `which gfortran` on your terminal. If it is not the the arm64 version that you installed, then you need to change the path of the default gfortran to the arm64. I'm adding a few words about this in the documentation. I'll be back to this thread to address the comments about the Rosetta installation.
