### 2024-11-25 c013fea
Update macos_arm64_install.rst
Clarify instructions in disk image section
