### 2024-11-25 7fc67e0
Update macos_arm64_install.rst
