### 2024-11-25 lorenaacuna/GASTLI#5 @lorenaacuna
I address the Rosetta installation questions:

> I am currently going through the installation with Rosetta for a M1/M2/M3 Mac and have a couple of questions and comments:
>
> 1. Twice in the installation instructions, you instruct to "cut and paste" -> do you really mean cut, or do you mean copy and paste? If it's indeed cut, it might be worth to clarify in parentheses `(copy and remove)`, if you don't actually mean cut, then replace the word "cut" with "copy". I have to admit that as someone who has never used Rosetta before, it is unclear which of the two is needed.
> The two lines are:
>
> * https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L70
> * https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L103

We mean cut, i.e copy and remove. I clarified this in the documentation.

>
> 2. [Here](https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L51), you say you'll assume people already have anaconda installed in their main terminal. Since most people are very confused between anaconda, miniconda and conda, here again it would be good to specify something like: "[...] you already installed `anaconda` or `miniconda`  in your M1/M2/M3 terminal [...]".

Ok, added to the documentation.

> 3. Claim: https://github.com/lorenaacuna/GASTLI/blob/de99c5ca4851d1b54d6cdf155bb1ce562b02cad2/docs/source/macos_arm64_install.rst?plain=1#L129
>     - but how will I actually know? Might be worth adding half a sentence describing how exactly the user will know.

It is displayed in the terminal when you open it. When I open my iTerm_Rosetta, this is the first message that is shown:

```
Last login: Mon Nov 25 10:04:04 on ttys000
Running on Rosetta using miniconda3

The default interactive shell is now zsh.
To update your account to use zsh, please run `chsh -s /bin/zsh`.
For more details, please visit https://support.apple.com/kb/HT208050.
(base) mac-n-176:~ acuna$

```

I added this to the documentation.

> 4. The most important comment is, actually, that when I ran `softwareupdate --install-rosetta`, I get the following output - note the error message in the second to last line before it claims it finished successfully:
>
> ```shell
> maintenncesMBP6:~ ilaginja$ softwareupdate --install-rosetta
> I have read and agree to the terms of the software license agreement. A list of Apple SLAs may be found here: http://www.apple.com/legal/sla/
> Type A and press return to agree: A
> 2024-11-24 20:33:44.639 softwareupdate[39894:12294479] Package Authoring Error: 002-57052: Package reference com.apple.pkg.RosettaUpdateAuto is missing installKBytes attribute
> Install of Rosetta 2 finished successfully
> ```
>
> The result is that I do not have the iTerm app on my machine. I tried googling this but wasn't able to understand what is happening. Would you have any clues?

I have Rosetta installed in the terminal in my OS Monterey M2 chip and I never had this message. Not having the iTerm app in the directory "Applications" of your Mac should not be a problem (I don't have it either myself). Use the default terminal app instead. In my case it is in Applications > Utilities, and it is named "Terminal". If you have the same, follow the instructions in the documentation with "Terminal" instead of "iTerm".

Then for the installation of Rosetta, you may already have Rosetta 2 installed in your system if you tried to run an Intel app before (it is downloaded automatically, see: https://support.apple.com/en-us/102527). If that's the case, you should be able to see the option "Open using Rosetta" under "Get Info" when your right click on the Terminal app.

Let me know if that works.
