### 2024-11-25 #7288 @lorenaacuna
@JBorrow I'm just waiting until @ivalaginja succesfully installs GASTLI in their system to close issue #5.

Can we send a kind reminder to the second reviewer (@MartianColonist) to send his comments soon?
