### 2024-12-01 #7288 @MartianColonist
Hi @lorenaacuna,

This first part of my review offers feedback on the software paper. The paper overall is very good and will make a great addition to JOSS (well done!). I hope my feedback and suggestions on the paper below will help improve its clarity.

**Feedback on the software paper**

_General comments_:

- There are a few missing sections commonly seen in JOSS papers:

1. **Similar Tools** (the [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html) requires "A list of key references, including to other software addressing related needs"): all this section needs is links to the code repositories of similar open source codes, such as MESA and MAGRATHEA.
2. **Documentation**: a single-sentence section pointing the user to the GASTLI documentation.
3. **Future Developments** (optional): it would be nice to mention any planned future developments / new applications of GASTLI after v1.0. For example, are there plans to support other atmospheric model grids in the future (besides pRT)?

- The Summary section is quite light on references to the exoplanet and interior modelling literature. I've noted below a few specific lines, but I do feel the first paragraph, in particular, needs additional citations to contextualise the field for beginners to GASTLI (and interior structure models more broadly).
- From a quick look at the [Exoplanet Modeling and Analysis Center](https://emac.gsfc.nasa.gov/?category=cdda4cb1-f016-4a65-b93f-79a28595c704&sort=date)), there may be a few other open source interior structure models (e.g., VPLanet and HARDCORE). In the Statement of Need / Similar Tools, it could be helpful to differentiate between open source packages that interpolate / fit existing interior model grids (e.g. plaNETic and SMINT) vs. full interior structure calculation codes like GASTLI.
- A Figure illustrating use cases of GASTLI would be a great addition that would significantly enhance the paper. For example, a schematic of a planet cross section with the GASTLI logo and text annotations stating applications. Such a Figure would tell a prospective user what problems GATLI can solve and the inputs and outputs of the code.
- The paper doesn't mention that GASTLI can be used to perform interior retrievals, but the code documentation highlights how to use GASTLI for this application. It would be nice to advertise this important feature in the paper.

_Line-by-line specific feedback_:

- Line 12: "Their bulk composition..." 'Their' is ambiguous in this sentence, as it could be referring to hot Jupiters (from the previous sentence) or gas giant exoplanets more broadly (from the paragraph subject). I'd presume the latter, but it would be good to update the wording to be specific. E.g. "The bulk composition of gaseous giant exoplanets..."
- Line 13-14: It would be helpful to provide a few more citations for how bulk compositions inform planet formation pathways and planet-star interactions.
- Line 16: please provide an example reference for the wide variety of mass fractions across the exoplanet population (e.g. a review paper).
- Lines 19-21: please include citations for at least one paper that uses inverse methods to infer the bulk composition of giant exoplanets.
- Lines 22-24: please provide an example citation for a paper/review discussing modelling inputs required for interior structure models.
- Line 31: typo "EOS" -> "EOSs".
- Line 41: 'retrieval' can be jargon for some readers, so I would recommend something like "Bayesian inverse fitting of mass and radius data ('retrieval')".
- Line 45: "The compositions of the envelope are limited to low metal content" - please clarify in this sentence if this limitation refers to just MESA or both codes mentioned (MESA and MAGRATHEA). Also, how low is 'low' here? Given Neptune is roughly 80x solar, where is the limit?
-  Line 49: "FORTRAN inlists" - I haven't encountered the term 'inlist' before, so perhaps this is specialised terminology for MESA? Is there an alternative more general term?
- Lines 49-51: "The computational time of one simple forward model is at least 1 minute. This renders MESA..." - I recommend merging these two sentences, since it isn't clear from the first alone that the quoted runtime is referring to MESA.
- Line 53: "computationally fast" - since a 1 minute runtime was mentioned for MESA, please also provide a rough forward model runtime for MAGRATHEA.
- Line 55: typo "type of planets" -> "types of planet".
- Line 56: please provide an example citation for the need to use non-ideal high-pressure H/He envelopes to model sub-Neptune interiors.
- Line 57: This section is very well-written and comprehensively referenced, nicely done.

Great work @lorenaacuna on the paper!

I am currently working through the code documentation and verifying GASTLI on my Linux machine. I will shortly provide further feedback on this in a separate comment and/or issues in the GASTLI GitHub repository.
