### 2024-12-01 lorenaacuna/GASTLI#5 @ivalaginja
> I have Rosetta installed in the terminal in my OS Monterey M2 chip and I never had this message. Not having the iTerm app in the directory "Applications" of your Mac should not be a problem (I don't have it either myself). Use the default terminal app instead. In my case it is in Applications > Utilities, and it is named "Terminal". If you have the same, follow the instructions in the documentation with "Terminal" instead of "iTerm".
>
> Then for the installation of Rosetta, you may already have Rosetta 2 installed in your system if you tried to run an Intel app before (it is downloaded automatically, see: https://support.apple.com/en-us/102527). If that's the case, you should be able to see the option "Open using Rosetta" under "Get Info" when your right click on the Terminal app.
>
> Let me know if that works.

This worked, thanks for the guidance. Maybe you could put this information in a "Warning" or "Note" box in the documentation too? Also, I first had to close all my open Terminal windows to make this work.

For the docs, I would urge you to remove the note about the conda env for gastli being optional which you have at the end of the Rosetta installation section. You go to great lengths having people install conda, so imho they should also use it. It is a good coding practice anyway so let's endorse it :)

That being said, I would like to see a clarification on the Installation page about the sequence of things. YOu currently have: "If you do not have either of the first two requirements, we offer an installation guide for different OS below: [...]"

Instead I would suggest to give very clear instructions for installation:
"To get everything installed, first make sure to follow the Fortran installation process appropriate for your OS, then proceed with the GASTLI package installation." -> I would also make a title for the Fortran installation section here on the same level like the title "Installation of GASTLI’s Python package". This way it is clear that these are two separate steps that need to be done in sequence to get everything done.

I will report back with my further tests.
