### 2024-12-01 lorenaacuna/GASTLI#6 @ivalaginja
Performance verification
For https://github.com/openjournals/joss-reviews/issues/7288.

Hi @lorenaacuna, in your paper you claim: "One single computation of the interior module (function `calc_radius` in the `GASTLI.int_planet` class) for the default precision and number of iterations takes 1.89 seconds in an Apple M2 CPU, with parallelization on 8 cores (four cores operating at 3.49 GHz and the other four cores operating at 2.42 GHz)."

Would you be able to provide the way to run/test this? I am trying to identify the code snippet I would have to time.
