### 2024-12-02 lorenaacuna/GASTLI#5 @lorenaacuna
> This worked, thanks for the guidance. Maybe you could put this information in a "Warning" or "Note" box in the documentation too? Also, I had to quite my Terminal to make this work which is logical to me once I realized you cannot run both architectures in separate windows at the same time apparently.
>
> For the docs, I would urge you to remove the note about the conda env for gastli being optional which you have at the end of the Rosetta installation section. You go to great lengths having people install conda, so imho they should also use it. It is a good coding practice anyway so let's endorse it :)
>
> That being said, I would like to see a clarification on the Installation page about the sequence of things. YOu currently have: "If you do not have either of the first two requirements, we offer an installation guide for different OS below: [...]"
>
> Instead I would suggest to give very clear instructions for installation:
> "To get everything installed, first make sure to follow the Fortran installation process appropriate for your OS, then proceed with the GASTLI package installation." -> I would also make a title for the Fortran installation section here on the same level like the title "Installation of GASTLI’s Python package". This way it is clear that these are two separate steps that need to be done in sequence to get everything done.
>

I'm adding these suggestions to the documentations now. I'm glad you could make everything work with Rosetta
