### 2024-12-02 lorenaacuna/GASTLI#5 @lorenaacuna
> Everything seems to run as expected. Do you have an idea though why the pytests don't run when I call them locally?
>
> <img alt="Screenshot 2024-12-01 at 21 33 27" width="668" src="https://private-user-images.githubusercontent.com/29508965/391363799-655c171e-cb76-40da-8160-e0b384d1bac0.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzMxNTYwMjAsIm5iZiI6MTczMzE1NTcyMCwicGF0aCI6Ii8yOTUwODk2NS8zOTEzNjM3OTktNjU1YzE3MWUtY2I3Ni00MGRhLTgxNjAtZTBiMzg0ZDFiYWMwLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDEyMDIlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMjAyVDE2MDg0MFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWEyYjFlNzdjYjk1ZjU2YjJmMDA0Mzc2N2FiZGM4NTQyN2QyZDJjNzcyZWQ1ZGI1MjBlNjkyYjUzM2M2MWJkMGQmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.UVOs0C4XhqC7-SijF9Q3JCKxiKkWRi0V3qa1FJPRimw">

It works for me. The exact steps I followed are:

1. Clone the repository with git. From your terminal: `git clone https://github.com/lorenaacuna/GASTLI`
2. Create a new virtual environment (or conda, whichever your prefer). In my case: `virtualenv GASTLI_forIvatest`
3. Activate the environment: `source GASTLI_forIvatest/bin/activate`
4. Let's check what packages are installed by default in this new environment: `pip list`. The output is:
```
Package    Version
---------- -------
pip        24.3.1
setuptools 75.3.0
wheel      0.44.0
```

5. Go to GASTLI main folder (`cd GASTLI`). If you run `ls` from this directory, you should see the meson.build and pyproject.toml files. I run `pip install .` to install GASTLI and its dependencies, including pytest. If you run `pip list` again after installing GASTLI, this should be the output:
```
Package         Version
--------------- -----------
contourpy       1.3.1
cycler          0.12.1
exceptiongroup  1.2.2
fonttools       4.55.0
gastli          0.9.1
h5py            3.12.1
iniconfig       2.0.0
kiwisolver      1.4.7
matplotlib      3.9.3
numpy           2.1.3
packaging       24.2
pandas          2.2.3
pillow          11.0.0
pip             24.3.1
pluggy          1.5.0
pyparsing       3.2.0
pytest          8.3.4
python-dateutil 2.9.0.post0
pytz            2024.2
scipy           1.14.1
setuptools      75.3.0
six             1.16.0
tomli           2.2.1
tzdata          2024.2
wheel           0.44.0

```
GASTLI and pytest are in the list.
6. Go to the the test directory (`cd test`) and run `pytest`. The output for me was:
```
(GASTLI_forIvatest) (base) mac-n-176:test acuna$ pytest
========================================================================= test session starts ==========================================================================
platform darwin -- Python 3.10.14, pytest-8.3.4, pluggy-1.5.0
rootdir: /Users/acuna/Desktop/MPIA2024/Code dev/JOSS GASTLI/Review/ivas test/GASTLI
configfile: pyproject.toml
collected 4 items

test_Coupling.py .                                                                                                                                               [ 25%]
test_GASTLI.py .                                                                                                                                                 [ 50%]
test_Thermal_evolution.py .                                                                                                                                      [ 75%]
test_atm_models_interp.py .                                                                                                                                      [100%]

==================================================================== 4 passed in 109.43s (0:01:49) =====================================================================
 [i] Allowed maximum number of iterations reached.

 [i] Convergence reached.
 [i] Convergence reached.
 [i] Convergence reached.
 [i] Allowed maximum number of iterations reached.

```

It takes a while for all tests to run (almost 2 minutes) but it run without interruptions.

I'm going to close this issue since we resolved the installation problems. If you still have problems with pytest and testing it, please open a new issue
