### 2024-12-02 lorenaacuna/GASTLI#7 @ivalaginja
Pytests don't run locally
For https://github.com/openjournals/joss-reviews/issues/7288.

@lorenaacuna unfortunately my issue with the pytests persists.

I have a conda environment that contains both `gastli` and its dependencies, as well as `pytest`. However, when I cd into the test directory of the GASTLI repo and run `pytest`, nothing happens.

![Screenshot 2024-12-02 at 17 44 59](https://github.com/user-attachments/assets/7dfc2b51-6f4b-4d8a-af9f-e20733eb33aa)
