### 2024-12-02 lorenaacuna/GASTLI#7 @lorenaacuna
Hi @ivalaginja, the only difference I notice between your terminal output and mine is the line
`plugins: anyio-4.6.2.post1`. I've made a quick Google search and anyIO is a new package that provides a plugin for testing with pytest. This is the first time I hear of anyIO, so none of the configuration files in GASTLI uses this package. Maybe it is your pytest/Python configuration that defaults to using the plugin. In any case, you can try fixing it by installing anyIO with `pip install anyio` in the same environment where you have GASTLI and pytest installed: https://anyio.readthedocs.io/en/stable/basics.html.

After installing anyIO, try running `pytest` from the test directory again. Let me know if that works.
