### 2024-12-02 lorenaacuna/GASTLI#7 @ivalaginja
I don't think I would have made any global pytest settings like this on my machine. In any case, if I try to resintall wih `pip install anyio`, it just tells me that this requirement is already satisfied and I get the same result when trying to run the pytests :/
