### 2024-12-02 lorenaacuna/GASTLI#7 @lorenaacuna
Can you try uninstalling it (`pip uninstall anyio` or `python -m pip uninstall anyio`) and run the test again?
