### 2024-12-03 c88ef1e
Update macos_arm64_install.rst
Added clarification about iTerm/Terminal for Rosetta installation
