### 2024-12-03 e8a4e2a
Update macos_arm64_install.rst
Remove "optional" conda/venv environment for Rosetta installation
