### 2024-12-03 de70b52
Update installation.rst
Fortran installation clarifications
