### 2024-12-03 34a317f
Update macos_arm64_install.rst
Added "recommended" to Rosetta instructions
