### 2024-12-03 lorenaacuna/GASTLI#7 @ivalaginja
Even after uninstalling I still get the same result...
```bash
platform darwin -- Python 3.10.15, pytest-8.3.4, pluggy-1.5.0
rootdir: /Users/ilaginja/repos/GASTLI
configfile: pyproject.toml
collecting ... (GASTLI_env) macbookiv:test ilaginja$
```
