### 2024-12-03 9fdb784
Update test_GASTLI.py
Added timing to interior module test for JOSS paper
