### 2024-12-03 lorenaacuna/GASTLI#6 @lorenaacuna
> Would you be able to provide the way to run/test this? I am trying to identify the code snippet I would have to time.

Hi @ivalaginja, the code snippet is the script `test_GASTLI.py` in the `test/` directory of the repository. I've added the commands to time the interior module function. If you run this script with `pytest` or with `python3 test_GASTLI.py` the output should be:

```
Running interior structure model
 [i] Allowed maximum number of iterations reached.

Calculation time [s] =  1.8798940181732178

```
