### 2024-12-07 lorenaacuna/GASTLI#7 @ivalaginja
I figured it out, the reason the pytests wouldn't run is because I wasn't aware that I needed to edit each test file manually to include my own path to files in them. Once that was done, the tests ran fine.

However, this is really not a good practice, so we will have to find a solution for this. A suite of pytests should run out of the box, with maybe a centralized configuration needing prior updates, but definitely not each individual test script. The latter is very intransparent to users and massively prone to "stupid" mistakes like just forgetting a path somewhere. It also doesn't help that the variable that holds the path to the files is currently not called the same everywhere, so it's impossible to search the project for all locations where the path would need updating. I already noticed this in the quickstart tutorials in the documentation.

Moreover, it is currently necessary to have a trailing slash at the end of the path to files, otherwise it is not joined correctly in the core of the gastli package. This should be fixed programmatically, otherwise it is likely that there will be both confusion and also OS-specific errors appearing.

To fix this, I would like to motivate you to do the following two things:
- Change the internal handling of the input file path. It should accept both a path with and without a trailing slash. This is easily achievable with the `os` package, using functions like `os.join()`, which adds a required slash dynamically if it is not present in the input path.
- Find a way to set up the local path to the input files only once on a user's local machine. There's a lot of fancy solutions out there with configuration files of different formats, but I think in your case, you could simply use an environment variable. And then you include instructions for the setup of that environment variable in your installation instructions, in the part where you tell people to download the data. Then, you replace each place in your code where the path should be put in with simply reading that environment variable, especially the tests and the quickstart tutorials.
