### 2024-12-14 lorenaacuna/GASTLI#10 @MartianColonist
Error: No such file or directory: '/tmp/pip-build-env-mmenj3lc/normal/bin/ninja' (JOSS review)
Hi @lorenaacuna,

After freshly installing GASTLI (by cloning the repository, making a new conda environment, and using `pip install .`), when I import the package via `import gastli.Coupling as cpl` I recieve the following error:

```
FileNotFoundError                         Traceback (most recent call last)
Cell In[1], [line 2](vscode-notebook-cell:?execution_count=1&line=2)
      [1](vscode-notebook-cell:?execution_count=1&line=1) # Import coupling module
----> [2](vscode-notebook-cell:?execution_count=1&line=2) import gastli.Coupling as cpl
      [3](vscode-notebook-cell:?execution_count=1&line=3) # Other Python modules
      [4](vscode-notebook-cell:?execution_count=1&line=4) import numpy as np

.
.
.

FileNotFoundError: [Errno 2] No such file or directory: '/tmp/pip-build-env-mmenj3lc/normal/bin/ninja'
```

Perhaps there is a missing dependency not specified in pyproject.toml?
