### 2024-12-14 #7288 @MartianColonist
Hi @lorenaacuna,

I've raised four issues on the GASTLI repository. They are mostly minor issues that cover suggestions to improve the installation instructions and documentation (thanks to @ivalaginja for your previous helpful suggestions!).

For the final part of my review, I am trying to run GASTLI to verify the functionality.

I have installed GASTLI following the instructions, but I received an error message upon importing the package ([issue #10](https://github.com/lorenaacuna/GASTLI/issues/10)). Please let me know once you have a solution to this, and then I'll continue running the tutorials.
