### 2024-12-16 9ea2b72
Update of input data path
Update: input data path is now handled with the environment variable "GASTLI_input_data_path"
