### 2024-12-16 9bcce9a
Update: input data path
Update: input data path is now handled as a environment variable
