+++ Examples/HAT-P-26b_thermal_solve_tutorial.py
@@ -8,12 +8,8 @@ import matplotlib.pyplot as plt
 from scipy import interpolate
 import pandas as pd
 
-# Path to input files
-# Dont forget the "/" at the end of the string
-path_input = "/Users/acuna/Desktop/gastli_input_data/"
-
 # Create thermal evolution class
-my_therm_obj = therm.thermal_evolution(path_to_file=path_input)
+my_therm_obj = therm.thermal_evolution()
 
 # Read in data saved in step 1
 data = pd.read_csv('thermal_sequence_HATP26b_CMF50_20xsolar.dat', sep='\s+',header=None,skiprows=1)
