+++ Examples/psurf_tutorial.py
@@ -12,8 +12,7 @@ import matplotlib.pyplot as plt
 
 # Create coupling class
 # Time initialization time
-my_coupling = cpl.coupling(path_to_file="/Users/acuna/Desktop/gastli_input_data/",\
-                           j_max=99, pow_law_formass=0.31)
+my_coupling = cpl.coupling(j_max=99, pow_law_formass=0.31)
 
 
 
