### 2024-12-16 b888ec0
Update: input data path
Update: input data path is now handled as a environment variable
