### 2024-12-16 26e468f
Update: version number
Update version 0.9.3 in meson.build and .toml
