### 2024-12-16 55878b1
Update CHANGELOG.rst
