+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 .. :changelog:
 
+0.9.3 (16 December 2024)
+~~~~~~~~~~~~~~~~~~
+
+- Update: input data path is handled as a environment variable.
+Suggested by JOSS reviewer.
+
+
 0.9.0 (5 July 2024)
 ~~~~~~~~~~~~~~~~~~
 
