### 2024-12-16 9d0a310
Update CHANGELOG.rst
