+++ CHANGELOG.rst
@@ -3,8 +3,7 @@
 0.9.3 (16 December 2024)
 ~~~~~~~~~~~~~~~~~~
 
-- Update: input data path is handled as a environment variable.
-Suggested by JOSS reviewer.
+- Update: input data path is handled as a environment variable. Suggested by JOSS reviewer.
 
 
 0.9.0 (5 July 2024)
