### 2024-12-16 0415be5
Update installation.rst
Update: added to documentation the instructions of how to set the input data path
