### 2024-12-16 e08f0b9
Update int_atm_object.rst
Removed path_to_input variable from documentation after 0.9.3 release
