### 2024-12-16 d3790ba
Update int_atm_object.rst
