### 2024-12-16 fadfd76
Update thermal_evol.rst
Update: removed path_to_input variable from documentation after release 0.9.3
