### 2024-12-16 7b071e7
Update convergence.rst
Removed variable path_to_input from documentation after release 0.9.3
