+++ docs/source/convergence.rst
@@ -49,8 +49,7 @@ We recommend increasing ``j_max``, to no more than ``j_max = 99``. Sometimes, ev
    import numpy as np
    import matplotlib.pyplot as plt
    # Create coupling class
-   my_coupling = cpl.coupling(path_to_file="/Users/acuna/Desktop/gastli_input_data/",\
-                           j_max=99, pow_law_formass=0.31)
+   my_coupling = cpl.coupling(j_max=99, pow_law_formass=0.31)
    # Input for interior
    M_P = 50.
    # Internal and equilibrium temperatures
