### 2024-12-16 0a0d6b3
Update atm_grid.rst
Removed path_to_input variable from documentation after release 0.9.3
