### 2024-12-16 82a3cbb
Update retrieval.rst
Removed path_to_input variable from documentation after release 0.9.3
