### 2024-12-16 lorenaacuna/GASTLI#7 @lorenaacuna
@ivalaginja thanks so much for catching this and the suggestions! It totally makes sense to handle it this way for the pytests to work.

I modified the source code, examples and tests to handle the input path with an environment variable. I updated the documentation to include a subsection about how to define the environment variable in the installation page.

To check the new feature and that the pytests work without having to modify the scripts:
1. Install GASTLI's latest version, 0.9.3 (`pip install gastli`, or `pip install gastli=0.9.3`) -  this is the one that handles the path with an environment variable.
2. Follow the installation page in the docs to define the env. variable `GASTLI_input_data_path`
3. Run the pytests as usual

Please let us know if the pytests work so I can close this issue (#7) and issue #6.
