### 2024-12-16 #7288 @lorenaacuna
> Editorial question for @JBorrow:
>
> I noticed in reviewing the software paper that there is another paper on GASTLI with a similar title, also advertising the package as an open source code, published in Astronomy & Astrophysics ([GASTLI: An open-source coupled interior-atmosphere model to unveil gas-giant composition](https://www.aanda.org/articles/aa/full_html/2024/08/aa50559-24/aa50559-24.html)). The main difference between this JOSS submission and the published methods paper is the description of the code classes and runtime in the JOSS submission.
>
> I presume these papers are sufficiently distinct that everything is good on JOSS's end?
>
> Also, is there any way to link the Astronomy & Astrophysics paper to this JOSS submission (in the same way an AAS Journal paired paper can appear in the margin of a JOSS paper)?

Thanks @MartianColonist for bringing this up.

We submitted a science paper to A&A before the software paper to JOSS for two reasons:

1. To make sure that the methodology - particularly the new implementation of H/He layer - and the scientific validations (comparison to Solar System planets and mass-radius relations from previous works) were peer-reviewed.

2. To highlight new scientific results that extend beyond the scope of a software paper. These include the intercomparison of interior models, illustrating how different assumptions influence the inferred metal mass fraction, as well as a retrieval analysis of HAT-P-26 b.

We believed these results would fall outside the intended scope of a JOSS software paper, which focuses more on the technical aspects, such as the description of code classes and runtime. Thus, we decided to submit the papers separately.

@JBorrow Please let us know if further clarification is needed.
