### 2024-12-16 #7288 @lorenaacuna
@MartianColonist thanks a lot for the review! I'm going to start addressing first your comments on the manuscript this week. After that I'll move to the docs and functionality suggestions/issues, but I may finish addressing those after the New Year. Thanks a lot for your patience.
