### 2024-12-17 lorenaacuna/GASTLI#7 @ivalaginja
Tests pass now without needing to adjust paths! This issue can be closed, however, please hold off on closing #6 for now.
