### 2024-12-22 #7288 @ivalaginja
@JBorrow @lorenaacuna I finished with my review and am happy with the state of the repo and paper. I think it will make a great contribution to JOSS. I recommend acceptance. Happy holidays!
