### 2025-01-02 #7288 @warrickball
Hi @MartianColonist — I apologise that it's taken so long to get some editorial consensus but to answer [your questions above](https://github.com/openjournals/joss-reviews/issues/7288#issuecomment-2509583510):

> I presume these papers are sufficiently distinct that everything is good on JOSS's end?

Yes. The editorial board agrees that the content of the A&A article distinguishes itself enough from the JOSS paper by focusing on the scientific background and applications, rather than the code itself.

> Also, is there any way to link the Astronomy & Astrophysics paper to this JOSS submission (in the same way an AAS Journal paired paper can appear in the margin of a JOSS paper)?

In short, no. That marker is on account of a specific agreement with AAS Journals (see the [announcement blog post](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing)), which we don't have with any other publishers.
