### 2025-01-29 abc0ca3
Added "Similar tools" section
