### 2025-01-29 #7288 @lorenaacuna
Hi @MartianColonist,

> Similar Tools (the [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html) requires "A list of key references, including to other software addressing related needs"): all this section needs is links to the code repositories of similar open source codes, such as MESA and MAGRATHEA.

> From a quick look at the [Exoplanet Modeling and Analysis Center](https://emac.gsfc.nasa.gov/?category=cdda4cb1-f016-4a65-b93f-79a28595c704&sort=date), there may be a few other open source interior structure models (e.g., VPLanet and HARDCORE). In the Statement of Need / Similar Tools, it could be helpful to differentiate between open source packages that interpolate / fit existing interior model grids (e.g. plaNETic and SMINT) vs. full interior structure calculation codes like GASTLI.

I've addressed these two points by adding a "Similar tools" section that contains a table summarising information about other codes. This includes (1) name, link and references, (2) whether it is an interior model package, or a trained neural network/grid interpolator, (3) planet type and components (H/He, water, rock, Fe) included in the modelling.

I'm addressing the next comments soon this week.
