### 2025-01-29 cfdd0ee
Addressed line-by-line comments from Ryan
