+++ docs/JOSS/paper.bib
@@ -1,4 +1,429 @@
 
+
+@INPROCEEDINGS{Marley07,
+       author = {{Marley}, M.~S. and {Fortney}, J. and {Seager}, S. and {Barman}, T.},
+        title = "{Atmospheres of Extrasolar Giant Planets}",
+     keywords = {Astrophysics},
+    booktitle = {Protostars and Planets V},
+         year = 2007,
+       editor = {{Reipurth}, Bo and {Jewitt}, David and {Keil}, Klaus},
+        month = jan,
+        pages = {733},
+          doi = {10.48550/arXiv.astro-ph/0602468},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0602468},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2007prpl.conf..733M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{Baraffe03,
+       author = {{Baraffe}, I. and {Chabrier}, G. and {Barman}, T.~S. and {Allard}, F. and {Hauschildt}, P.~H.},
+        title = "{Evolutionary models for cool brown dwarfs and extrasolar giant planets. The case of HD 209458}",
+      journal = {\aap},
+     keywords = {planetary systems, stars: brown dwarfs, stars: evolution, stars: individual (HD 209458), Astrophysics},
+         year = 2003,
+        month = may,
+       volume = {402},
+        pages = {701-712},
+          doi = {10.1051/0004-6361:20030252},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0302293},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2003A&A...402..701B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{Burrows97,
+       author = {{Burrows}, A. and {Sudarsky}, D. and {Sharp}, C. and {Marley}, M. and {Hubbard}, W.~B. and {Lunine}, J.~I. and {Guillot}, T. and {Saumon}, D. and {Freedman}, R.},
+        title = "{Advances in the Theory of Extrasolar Giant Planets and Brown Dwarfs}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics},
+         year = 1997,
+        month = jun,
+          eid = {astro-ph/9706080},
+        pages = {astro-ph/9706080},
+          doi = {10.48550/arXiv.astro-ph/9706080},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/9706080},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1997astro.ph..6080B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{Rogers23,
+       author = {{Rogers}, James G. and {Schlichting}, Hilke E. and {Owen}, James E.},
+        title = "{Conclusive Evidence for a Population of Water Worlds around M Dwarfs Remains Elusive}",
+      journal = {\apjl},
+     keywords = {Exoplanet atmospheric evolution, Exoplanet atmospheric composition, 2308, 2021, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2023,
+        month = apr,
+       volume = {947},
+       number = {1},
+          eid = {L19},
+        pages = {L19},
+          doi = {10.3847/2041-8213/acc86f},
+archivePrefix = {arXiv},
+       eprint = {2301.04321},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...947L..19R},
[…]
