+++ docs/JOSS/paper.md
@@ -37,14 +37,14 @@ bibliography: paper.bib
 
 # Summary
 
-Gaseous exoplanets are far more diverse in mass, density and temperature than their Solar System counterparts. These range from temperate sub-Neptunes whose masses and sizes are below that of the Solar System ice giants, to extremely irradiated, Jupiter-mass exoplanets known as hot Jupiters. Their bulk composition is key to understanding their formation and evolution pathways, including their dominant formation mechanism [@Thorngren16] and interactions between the irradiation of the host star and the planetary atmosphere [@OW13]. Their building blocks consist of rocks and water (metals), and hydrogen and helium (H/He), which are present in a wide variety of mass fractions in the exoplanet population. For exoplanets that contain H/He, the planetary radius (or density) not only depends on mass, but also on the temperature due to the distance from its host star, its age, the mass of the core, and the fraction of H/He in its envelope. The inference of bulk composition from observational data -- mass, radius, equilibrium temperature, age and atmospheric metallicity -- requires non-trivial numerical methods that solve the equations of planetary interior structure. Additionally, interior models need to incorporate the effects of processes occurring in the planet atmosphere, such as radiative transfer, and the thermodynamical properties of the different materials, provided by the equation of state (EOS).
+Gaseous exoplanets are far more diverse in mass, density and temperature than their Solar System counterparts. These range from temperate sub-Neptunes whose masses and sizes are below that of the Solar System ice giants, to extremely irradiated, Jupiter-mass exoplanets known as hot Jupiters. The bulk composition of sub-Neptunes and gas giants is key to understanding their formation and evolution pathways, including their dominant formation mechanism [@Thorngren16,@Teske19,@Ginzburg20,@Schneider21,@Mah_Bitsch23] and interactions between the irradiation of the host star and the planetary atmosphere [@OW13,@Acuna22,@Rogers24]. Their building blocks consist of rocks and water (metals), and hydrogen and helium (H/He), which are present in a wide variety of mass fractions in the exoplanet population [@ariel_int_rev]. For exoplanets that contain H/He, the planetary radius (or density) not only depends on mass, but also on the temperature due to the distance from its host star [@Burrows97], its age [@Baraffe03], the mass of the core [@Baraffe08], and the fraction of H/He in its envelope [@Rogers23]. The inference of bulk composition from observational data -- mass, radius, equilibrium temperature, age and atmospheric metallicity -- requires non-trivial numerical methods that solve the equations of planetary interior structure [@Miguel_rev], in addition to Bayesian methods to solve the inverse problem [@Dorn15,@Otegi20,@Acuna21,@Bloot23]. Additionally, interior models need to incorporate the effects of processes occurring in the planet atmosphere, such as radiative transfer [@Baraffe03,@Marley07], and the thermodynamical properties of the different materials, provided by the equation of state (EOS) [@HG23,@Howard23,@gastli_science].
 
[…]
