### 2025-01-29 bc43bb8
Corrections to .md and .bib files
