+++ docs/JOSS/paper.bib
@@ -119,7 +119,7 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{Miguel_rev,
-       author = {{Miguel}, Yamila and {Vazan}, Allona},
+       author = {{Miguel}, Y. and {Vazan}, A.},
         title = "{Interior and Evolution of the Giant Planets}",
       journal = {Remote Sensing},
      keywords = {giant planets interiors, giant planets evolution, planet formation, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
@@ -1201,7 +1201,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{HG23,
-       author = {{Howard}, Saburo and {Guillot}, Tristan},
+       author = {{Howard}, S. and {Guillot}, T.},
         title = "{Accounting for non-ideal mixing effects in the hydrogen-helium equation of state}",
       journal = {arXiv e-prints},
      keywords = {Astrophysics - Earth and Planetary Astrophysics},
