+++ docs/JOSS/paper.bib
@@ -324,6 +324,23 @@ archivePrefix = {arXiv},
 }
 
 
+@ARTICLE{Zeng19,
+       author = {{Zeng}, Li and {Jacobsen}, Stein B. and {Sasselov}, Dimitar D. and {Petaev}, Michail I. and {Vanderburg}, Andrew and {Lopez-Morales}, Mercedes and {Perez-Mercader}, Juan and {Mattsson}, Thomas R. and {Li}, Gongjie and {Heising}, Matthew Z. and {Bonomo}, Aldo S. and {Damasso}, Mario and {Berger}, Travis A. and {Cao}, Hao and {Levi}, Amit and {Wordsworth}, Robin D.},
+        title = "{Growth model interpretation of planet size distribution}",
+      journal = {Proceedings of the National Academy of Science},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics, Physics - Geophysics},
+         year = 2019,
+        month = may,
+       volume = {116},
+       number = {20},
+        pages = {9723-9728},
+          doi = {10.1073/pnas.1812905116},
+archivePrefix = {arXiv},
+       eprint = {1906.04253},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019PNAS..116.9723Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 @ARTICLE{Rogers24,
        author = {{Rogers}, James G. and {Owen}, James E. and {Schlichting}, Hilke E.},
@@ -481,30 +498,6 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2023A&A...675A..52C},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
-
-@ARTICLE{Agui21,
-       author = {{Aguichine}, Artyom and {Mousis}, Olivier and {Deleuil}, Magali and {Marcq}, Emmanuel},
-        title = "{Mass-Radius Relationships for Irradiated Ocean Planets}",
-      journal = {\apj},
-     keywords = {Exoplanets, Hydrosphere, Planetary interior, Planetary theory, Exoplanet astronomy, Exoplanet structure, Computational methods, 498, 770, 1248, 1258, 486, 495, 1965, Astrophysics - Earth and Planetary Astrophysics},
-         year = 2021,
-        month = jun,
-       volume = {914},
-       number = {2},
-          eid = {84},
-        pages = {84},
-          doi = {10.3847/1538-4357/abfa99},
-archivePrefix = {arXiv},
-       eprint = {2105.01102},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021ApJ...914...84A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-
 
 @ARTICLE{Zeng16,
        author = {{Zeng}, Li and {Sasselov}, Dimitar D. and {Jacobsen}, Stein B.},
