### 2025-01-29 741307f
Added summary figure to JOSS paper
