+++ docs/JOSS/paper.md
@@ -50,10 +50,12 @@ Interior structure codes are essential to carry out planet composition analyses 
 
 # The GASTLI Python package
 
-
 The interior model module is contained in the `GASTLI.int_planet` class. The interior structure equations for hydrostatic equilibrium, Gauss's theorem for gravitational acceleration, convection, and mass conservation are solved by combining integration by the trapezoidal rule with an iteration scheme over all profiles (pressure, temperature, gravity and density). In each iteration, the trapezoidal integrations are carried for each point in a 1D grid that represents the radius, from the center of the planet to the outermost boundary of the top layer. The density is calculated by evaluating the EOS for rock [@sesame; @Miguel22], water [@Mazevet19] and H/He [@Chabrier21; @HG23] at the pressure and temperature conditions of each point in the spatial grid. The iteration sequence is considered to have converged when the difference between iterations is less than a given precision that can be modified by the user. The recommended and default value for this precision is $10^{-5}$. The iteration scheme can also be stopped at a given number of iterations defined by the user, with a default of 30 maximum iterations, and a maximum possible value of 99 iterations. The boundary conditions to solve these equations include the top layer's pressure and temperature, which can be defined by the user. In addition to the profile convergence, the interior module ensures that the surface boundary conditions and the input mass are satisfied. A complete description of the interior module integration scheme can be found in @Brugger_phd_thesis, and has been widely used [@Brugger16; @Brugger17; @Mousis20; @Acuna21; @Aguichine21; @Acuna22; @Vivien22; @Acuna23]. The new feature of the interior module includes the optimization of several of its routines, such as the trapezoidal integration and the evaluation of the EOS's density tables, and the parallelization of the calculation of the interior structure profiles with OpenMP in FORTRAN. One single computation of the interior module (function `calc_radius` in the `GASTLI.int_planet` class) for the default precision and number of iterations takes 1.89 seconds in an Apple M2 CPU, with parallelization on 8 cores (four cores operating at 3.49 GHz and the other four cores operating at 2.42 GHz).
 
 The class `atm_models_interp` encloses the function that interpolates the atmospheric grid to obtain the temperature profile and interior boundary temperature. In addition, this class also provides the functions that calculate the atmospheric thickness and the atmosphere density profile. The latter is obtained by evaluating the EOS used by the interior module for H/He, and the AQUA water EOS [@Haldemann20]. The atmospheric grid class and the interior module class are then combined in the `Coupling` class. This class self-consistently couples the interior module and the atmosphere grids to calculate the boundary temperature of the outermost layer of the interior model. See @Acuna21 for a detailed description of the coupling algorithm, which has been used in previous work [@Aguichine21; @Acuna22; @Acuna23]. Finally, the class `Thermal_evolution` contains the functions necessary to evaluate a sequence of coupled interior-atmosphere models at different internal temperatures and solve the luminosity equation to obtain the internal temperatures and radii as a function of age [@Fortney13; @Poser19]. The complete validation of GASTLI against mass-radius relationships obtained with interior models used in previous work [@Fortney13; @Muller21] can be found in @gastli_science.
+
+![Schematic diagram of GASTLI's compositional structure, inputs and outputs. \label{fig:summary_fig}](GASTLI_summary_fig.pdf)
+
 
 # Similar tools
 
