### 2025-01-29 c6f0730
Added paragraph about retrievals
