+++ docs/JOSS/paper.bib
@@ -1,5 +1,64 @@
 
 
+
+@ARTICLE{emcee,
+       author = {{Foreman-Mackey}, Daniel and {Hogg}, David W. and {Lang}, Dustin and {Goodman}, Jonathan},
+        title = "{emcee: The MCMC Hammer}",
+      journal = {\pasp},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Statistics - Computation},
+         year = 2013,
+        month = mar,
+       volume = {125},
+       number = {925},
+        pages = {306},
+          doi = {10.1086/670067},
+archivePrefix = {arXiv},
+       eprint = {1202.3665},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013PASP..125..306F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+
+@ARTICLE{Mosegaard,
+       author = {{Mosegaard}, Klaus and {Tarantola}, Albert},
+        title = "{Monte Carlo sampling of solutions to inverse problems}",
+      journal = {\jgr},
+     keywords = {Mathematical Geophysics: Inverse theory},
+         year = 1995,
+        month = jul,
+       volume = {100},
+       number = {B7},
+        pages = {12,431-12,447},
+          doi = {10.1029/94JB03097},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1995JGR...10012431M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{nested_sampling,
+       author = {{Buchner}, Johannes},
+        title = "{Nested Sampling Methods}",
+      journal = {Statistics Surveys},
+     keywords = {Statistics - Computation, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2023,
+        month = jan,
+       volume = {17},
+        pages = {169-215},
+          doi = {10.1214/23-SS144},
+archivePrefix = {arXiv},
+       eprint = {2101.09675},
+ primaryClass = {stat.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023StSur..17..169B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+
 @INPROCEEDINGS{Marley07,
        author = {{Marley}, M.~S. and {Fortney}, J. and {Seager}, S. and {Barman}, T.},
         title = "{Atmospheres of Extrasolar Giant Planets}",
