+++ docs/JOSS/paper.md
@@ -54,7 +54,12 @@ The interior model module is contained in the `GASTLI.int_planet` class. The int
 
 The class `atm_models_interp` encloses the function that interpolates the atmospheric grid to obtain the temperature profile and interior boundary temperature. In addition, this class also provides the functions that calculate the atmospheric thickness and the atmosphere density profile. The latter is obtained by evaluating the EOS used by the interior module for H/He, and the AQUA water EOS [@Haldemann20]. The atmospheric grid class and the interior module class are then combined in the `Coupling` class. This class self-consistently couples the interior module and the atmosphere grids to calculate the boundary temperature of the outermost layer of the interior model. See @Acuna21 for a detailed description of the coupling algorithm, which has been used in previous work [@Aguichine21; @Acuna22; @Acuna23]. Finally, the class `Thermal_evolution` contains the functions necessary to evaluate a sequence of coupled interior-atmosphere models at different internal temperatures and solve the luminosity equation to obtain the internal temperatures and radii as a function of age [@Fortney13; @Poser19]. The complete validation of GASTLI against mass-radius relationships obtained with interior models used in previous work [@Fortney13; @Muller21] can be found in @gastli_science.
 
+
+
 ![Schematic diagram of GASTLI's compositional structure, inputs and outputs. \label{fig:summary_fig}](GASTLI_summary_fig.pdf)
+
+
+In an interior retrieval, we fit an interior composition model to data, particularly to mass, radius, age, atmospheric metallicity and/or internal temperature. This is usually done through Bayesian model fitting, such as Markov chain Monte Carlo [@Mosegaard] or nested sampling [@nested_sampling]. The likelihood function in interior retrievals is computed as a function of the squared residuals between the compositional model and the observed values of mass, radius, etc (see equations 6 and 14 in @Dorn15 and @Acuna21, respectively). GASTLI is fast enough to generate a grid of forward models that explores all possible parameter space for an exoplanet. Such a grid may contain thousands of thermal evolution tracks. For example, a grid of 4000 models can be computed in 12 hours, depending on the computer architecture and number of threads/cores available for parallelization, given that each model's computational time is 2-10 seconds. In GASTLI's documentation, we show an example of how to generate a grid and interpolate it to carry an interior retrieval. This example uses as sampler emcee [@emcee]. 
 
 
 # Similar tools
