### 2025-01-29 f0baf0f
Added a paragraph on atmospheric grids
