+++ docs/JOSS/paper.md
@@ -59,7 +59,9 @@ The class `atm_models_interp` encloses the function that interpolates the atmosp
 ![Schematic diagram of GASTLI's compositional structure, inputs and outputs. \label{fig:summary_fig}](GASTLI_summary_fig.pdf)
 
 
-In an interior retrieval, we fit an interior composition model to data, particularly to mass, radius, age, atmospheric metallicity and/or internal temperature. This is usually done through Bayesian model fitting, such as Markov chain Monte Carlo [@Mosegaard] or nested sampling [@nested_sampling]. The likelihood function in interior retrievals is computed as a function of the squared residuals between the compositional model and the observed values of mass, radius, etc (see equations 6 and 14 in @Dorn15 and @Acuna21, respectively). GASTLI is fast enough to generate a grid of forward models that explores all possible parameter space for an exoplanet. Such a grid may contain thousands of thermal evolution tracks. For example, a grid of 4000 models can be computed in 12 hours, depending on the computer architecture and number of threads/cores available for parallelization, given that each model's computational time is 2-10 seconds. In GASTLI's documentation, we show an example of how to generate a grid and interpolate it to carry an interior retrieval. This example uses as sampler emcee [@emcee]. 
+In an interior retrieval, we fit an interior composition model to data, particularly to mass, radius, age, atmospheric metallicity and/or internal temperature. This is usually done through Bayesian model fitting, such as Markov chain Monte Carlo [@Mosegaard] or nested sampling [@nested_sampling]. The likelihood function in interior retrievals is computed as a function of the squared residuals between the compositional model and the observed values of mass, radius, etc. The likelihood function use in interior retrievals can be found in equations 6 and 14 of @Dorn15 and @Acuna21, respectively. GASTLI can generate a grid of forward models that contains thousands of thermal evolution tracks and explores all possible parameter space for an exoplanet. For example, a grid of 4000 models can be computed in 12 hours, depending on the computer architecture and number of threads/cores available for parallelization, given that each model's computational time is 2-10 seconds. In GASTLI's documentation, we show an example of how to generate a grid and interpolate it to carry an interior retrieval. This example uses as sampler emcee [@emcee]. 
+
+Finally, GASTLI provides a default atmospheric grid for gas giants and sub-Neptunes with equilibrium temperatures under 1000 K, and atmospheric compositions from sub-solar to 250x solar (80% metal mass fraction). Our default grid assumes clear atmospheres in chemical equilibrium. If this grid is not appropriate for a particular target, the user may choose from a variety of model grids that are publicly available. These grids include clear, self-luminous atmospheres in equilibrium [@bobcat;@Linder18;samland17], clear, self-luminous in disequilibrium [@elfowl;@cholla], cloudy self-luminous [@diamondback;@lacy_burrows;Jorgensen24], cloudy irradiated gas giants [@charnay18], hot Jupiters [@Molliere15;@Roth24] and temperate sub-Neptunes [@Kempton23]. GASTLI's documentation includes an example of the grid file format and input.
 
 
 # Similar tools
