+++ docs/JOSS/paper.bib
@@ -1,4 +1,210 @@
 
+@ARTICLE{Kempton23,
+       author = {{Kempton}, Eliza M. -R. and {Lessard}, Madeline and {Malik}, Matej and {Rogers}, Leslie A. and {Futrowsky}, Kate E. and {Ih}, Jegug and {Marounina}, Nadejda and {Romero-Mirza}, Carlos E.},
+        title = "{Where are the Water Worlds?: Self-consistent Models of Water-rich Exoplanet Atmospheres}",
+      journal = {\apj},
+     keywords = {Exoplanet atmospheres, Exoplanet astronomy, Exoplanets, Mini Neptunes, Radiative transfer, 487, 486, 498, 1063, 1335, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2023,
+        month = aug,
+       volume = {953},
+       number = {1},
+          eid = {57},
+        pages = {57},
+          doi = {10.3847/1538-4357/ace10d},
+archivePrefix = {arXiv},
+       eprint = {2307.06508},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...953...57K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+
+@ARTICLE{Jorgensen24,
+       author = {{J{\o}rgensen}, Uffe G. and {Amadio}, Flavia and {Campos Estrada}, Beatriz and {M{\o}ller}, Kristian Holten and {Schneider}, Aaron D. and {Balduin}, Thorsten and {D'Alessandro}, Azzurra and {Symeonidou}, Eftychia and {Helling}, Christiane and {Nordlund}, {\r{A}}ke and {Woitke}, Peter},
+        title = "{A grid of self-consistent MSG (MARCS-StaticWeather-GGchem) cool stellar, sub-stellar, and exoplanetary model atmospheres}",
+      journal = {\aap},
+     keywords = {molecular data, opacity, stars: atmospheres, stars: late-type, stars: low-mass, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2024,
+        month = oct,
+       volume = {690},
+          eid = {A127},
+        pages = {A127},
+          doi = {10.1051/0004-6361/202450108},
+archivePrefix = {arXiv},
+       eprint = {2407.09397},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...690A.127J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{Roth24,
+       author = {{Roth}, Alexander and {Parmentier}, Vivien and {Hammond}, Mark},
+        title = "{Hot Jupiter diversity and the onset of TiO/VO revealed by a large grid of non-grey global circulation models}",
+      journal = {\mnras},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics},
+         year = 2024,
+        month = jun,
+       volume = {531},
+       number = {1},
+        pages = {1056-1083},
+          doi = {10.1093/mnras/stae984},
+archivePrefix = {arXiv},
+       eprint = {2404.09626},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024MNRAS.531.1056R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@ARTICLE{charnay18,
[…]
