+++ docs/JOSS/paper.bib
@@ -187,7 +187,7 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{bobcat,
-       author = {{Marley}, Mark S. and {Saumon}, Didier and {Visscher}, Channon and {Lupu}, Roxana and {Freedman}, Richard and {Morley}, Caroline and {Fortney}, Jonathan J. and {Seay}, Christopher and {Smith}, Adam J.~R.~W. and {Teal}, D.~J. and {Wang}, Ruoyan},
+       author = {{Marley}, M. S. and {Saumon}, Didier and {Visscher}, Channon and {Lupu}, Roxana and {Freedman}, Richard and {Morley}, Caroline and {Fortney}, Jonathan J. and {Seay}, Christopher and {Smith}, Adam J.~R.~W. and {Teal}, D.~J. and {Wang}, Ruoyan},
         title = "{The Sonora Brown Dwarf Atmosphere and Evolution Models. I. Model Description and Application to Cloudless Atmospheres in Rainout Chemical Equilibrium}",
       journal = {\apj},
      keywords = {Brown dwarfs, T dwarfs, L dwarfs, L subdwarfs, T subdwarfs, Y dwarfs, Stellar atmospheres, 185, 1679, 894, 896, 1680, 1827, 1584, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
@@ -266,7 +266,7 @@ archivePrefix = {arXiv},
 
 
 @INPROCEEDINGS{Marley07,
-       author = {{Marley}, M.~S. and {Fortney}, J. and {Seager}, S. and {Barman}, T.},
+       author = {{Marley}, M. S. and {Fortney}, J. and {Seager}, S. and {Barman}, T.},
         title = "{Atmospheres of Extrasolar Giant Planets}",
      keywords = {Astrophysics},
     booktitle = {Protostars and Planets V},
