### 2025-01-29 #7288 @lorenaacuna
@MartianColonist, I have addressed the rest of the points in [this comment](https://github.com/openjournals/joss-reviews/issues/7288#issuecomment-2509581553).

I'll continue with the documentation and testing issues soon.
