### 2025-01-29 #7288 @MartianColonist
Hi @lorenaacuna, excellent work; the paper looks exquisite!

I've updated my review to reflect the paper being complete.

All that remains now is verifying the install and functionality of GASTLI (per the GitHub issues).
