### 2025-01-30 88946b0
Added control parameter j_min
j_min: minimum number of iterations to let radius converge within interior module. Useful for low surface gravity cases. Default value is now j_max = 100, j_min = 15. j_max must always be below 999, and j_min < j_max.
