### 2025-02-17 cbecf1d
Logo rendering fixed
Changed line 26 in conf.py from .pdf to .png
