### 2025-02-17 605371b
Update installation instructions
Updated installations instructions following issue #9
