### 2025-02-17 ec910de
Changed suggested index
Updated changes to index.rst following comment 1 in issue #11
