### 2025-02-17 6666fe4
Update int_atm_object.rst
