### 2025-02-17 48e3e85
Update int_atm_object.rst
