### 2025-02-17 a96f8e7
Update int_atm_object.rst
