### 2025-02-17 399464b
Update int_atm_object.rst
