### 2025-02-17 6de88ec
Update int_atm_object.rst
