### 2025-02-17 d841937
Update int_atm_object.rst
