### 2025-02-17 a10f4e2
Update thermal_evol.rst
