### 2025-02-17 e6c213c
Update convergence.rst
