### 2025-02-17 07d897e
Update retrieval.rst
