### 2025-02-17 lorenaacuna/GASTLI#10 @lorenaacuna
@MartianColonist ninja seems to be already installed in most managers (conda, venv, etc). You are the first user to report not having it by default. I've opened a new branch called "build_dev", and added ninja to the dependencies in the .toml file:

```
[build-system]
build-backend = "mesonpy"
requires = [
    "meson-python >=0.15.0",
    "numpy >= 1.25.0",
    "ninja"
]
```

Please, try installing gastli again with this branch and see if it works. If it doesn't, it may have to do with the backend builder, meson, not being able to find the path of ninja in your installation:

See this issue: https://github.com/mesonbuild/meson/issues/10987
