### 2025-03-03 #7288 @MartianColonist
Thanks for the reminder! I'm taking a look at the last issues today. Hopefully I can install and run GASTLI now 🤞
