### 2025-03-03 lorenaacuna/GASTLI#10 @MartianColonist
Hi @lorenaacuna,

My install of GASTLI appears to have worked on "build_dev"!

In the "quickstart" guide, I get the following error:

> At line 328 of file ../src/fort_input.f95 (unit = 158)
> Fortran runtime error: Cannot open file '/home/ryanmacdonald/code/GASTLI/gastli_input_data/Input/Mazevet_water/EOS_water_mazevet.dat': No such file or directory

Inspecting the GASTLI input files, the problem appears to be a typo (the file in that folder is 'EOS_water_Mazevet.dat')
