### 2025-03-03 lorenaacuna/GASTLI#10 @MartianColonist
After changing the file name to fix the typo, I confirm that everything works in the quick start tutorial! 🎉
