### 2025-03-03 #7288 @MartianColonist
@JBorrow, I have now successfully installed and ran GASTLI, verifying the package's performance and functionality.

I am happy to recommend publication in JOSS.

Well done @lorenaacuna!
