### 2025-03-04 lorenaacuna/GASTLI#10 @lorenaacuna
Thanks @MartianColonist ! It seems Fortran is case sensitive in Linux, that's why that bug did not come up in other OSs. I've corrected the typo in the Fortran source code.
