+++ docs/source/conf.py
@@ -6,8 +6,8 @@ project = 'GASTLI'
 copyright = '2024, Acuña'
 author = 'Acuña'
 
-release = '0.1'
-version = '0.1.0'
+release = '1.0'
+version = '1.0.0'
 
 # -- General configuration
 
