### 2025-03-05 #7288 @JBorrow
Some minor comments on the text, before completion:

L92: The new feature**s** of the interior module include~s~ the ...
L119: The likelihood function use**d** in interior...
L126: This example uses `emcee` as a sampler OR This example uses `emcee` for sampling

Once these minor text changes have been made, we can proceed with passing this up to the editor in chief for final approval. I will be on vacation for the next 12 days, so please do not expect a response from me before then.
