+++ docs/JOSS/paper.md
@@ -50,7 +50,7 @@ Interior structure codes are essential to carry out planet composition analyses 
 
 # The GASTLI Python package
 
-The interior model module is contained in the `GASTLI.int_planet` class. The interior structure equations for hydrostatic equilibrium, Gauss's theorem for gravitational acceleration, convection, and mass conservation are solved by combining integration by the trapezoidal rule with an iteration scheme over all profiles (pressure, temperature, gravity and density). In each iteration, the trapezoidal integrations are carried for each point in a 1D grid that represents the radius, from the center of the planet to the outermost boundary of the top layer. The density is calculated by evaluating the EOS for rock [@sesame; @Miguel22], water [@Mazevet19] and H/He [@Chabrier21; @HG23] at the pressure and temperature conditions of each point in the spatial grid. The iteration sequence is considered to have converged when the difference between iterations is less than a given precision that can be modified by the user. The recommended and default value for this precision is $10^{-5}$. The iteration scheme can also be stopped at a given number of iterations defined by the user, with a default of 30 maximum iterations, and a maximum possible value of 99 iterations. The boundary conditions to solve these equations include the top layer's pressure and temperature, which can be defined by the user. In addition to the profile convergence, the interior module ensures that the surface boundary conditions and the input mass are satisfied. A complete description of the interior module integration scheme can be found in @Brugger_phd_thesis, and has been widely used [@Brugger16; @Brugger17; @Mousis20; @Acuna21; @Aguichine21; @Acuna22; @Vivien22; @Acuna23]. The new feature of the interior module includes the optimization of several of its routines, such as the trapezoidal integration and the evaluation of the EOS's density tables, and the parallelization of the calculation of the interior structure profiles with OpenMP in FORTRAN. One single computation of the interior module (function `calc_radius` in the `GASTLI.int_planet` class) for the default precision and number of iterations takes 1.89 seconds in an Apple M2 CPU, with parallelization on 8 cores (four cores operating at 3.49 GHz and the other four cores operating at 2.42 GHz).
[…]
