### 2025-03-18 #7288 @JBorrow
Thanks @lorenaacuna! Would it be possible to make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper? It seems that the title is different.
