+++ docs/JOSS/paper.bib
@@ -284,23 +284,6 @@ archivePrefix = {arXiv},
 
 
 
-@ARTICLE{Baraffe03,
-       author = {{Baraffe}, I. and {Chabrier}, G. and {Barman}, T.~S. and {Allard}, F. and {Hauschildt}, P.~H.},
-        title = "{Evolutionary models for cool brown dwarfs and extrasolar giant planets. The case of HD 209458}",
-      journal = {\aap},
-     keywords = {planetary systems, stars: brown dwarfs, stars: evolution, stars: individual (HD 209458), Astrophysics},
-         year = 2003,
-        month = may,
-       volume = {402},
-        pages = {701-712},
-          doi = {10.1051/0004-6361:20030252},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0302293},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2003A&A...402..701B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
 
 
 @ARTICLE{Burrows97,
