### 2025-03-20 14fd595
Initial editorial tweaks for JOSS paper
