+++ docs/JOSS/paper.bib
@@ -1,8 +1,8 @@
 
 @ARTICLE{Kempton23,
        author = {{Kempton}, Eliza M. -R. and {Lessard}, Madeline and {Malik}, Matej and {Rogers}, Leslie A. and {Futrowsky}, Kate E. and {Ih}, Jegug and {Marounina}, Nadejda and {Romero-Mirza}, Carlos E.},
-        title = "{Where are the Water Worlds?: Self-consistent Models of Water-rich Exoplanet Atmospheres}",
-      journal = {\apj},
+        title = {Where are the Water Worlds?: Self-consistent Models of Water-rich Exoplanet Atmospheres},
+      journal = {The Astrophysical Journal},
      keywords = {Exoplanet atmospheres, Exoplanet astronomy, Exoplanets, Mini Neptunes, Radiative transfer, 487, 486, 498, 1063, 1335, Astrophysics - Earth and Planetary Astrophysics},
          year = 2023,
         month = aug,
@@ -24,7 +24,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Jorgensen24,
        author = {{J{\o}rgensen}, Uffe G. and {Amadio}, Flavia and {Campos Estrada}, Beatriz and {M{\o}ller}, Kristian Holten and {Schneider}, Aaron D. and {Balduin}, Thorsten and {D'Alessandro}, Azzurra and {Symeonidou}, Eftychia and {Helling}, Christiane and {Nordlund}, {\r{A}}ke and {Woitke}, Peter},
         title = "{A grid of self-consistent MSG (MARCS-StaticWeather-GGchem) cool stellar, sub-stellar, and exoplanetary model atmospheres}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {molecular data, opacity, stars: atmospheres, stars: late-type, stars: low-mass, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2024,
         month = oct,
@@ -44,7 +44,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Roth24,
        author = {{Roth}, Alexander and {Parmentier}, Vivien and {Hammond}, Mark},
         title = "{Hot Jupiter diversity and the onset of TiO/VO revealed by a large grid of non-grey global circulation models}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {Astrophysics - Earth and Planetary Astrophysics},
          year = 2024,
         month = jun,
@@ -63,8 +63,8 @@ archivePrefix = {arXiv},
 
 @ARTICLE{charnay18,
        author = {{Charnay}, B. and {B{\'e}zard}, B. and {Baudino}, J. -L. and {Bonnefoy}, M. and {Boccaletti}, A. and {Galicher}, R.},
-        title = "{A Self-consistent Cloud Model for Brown Dwarfs and Young Giant Exoplanets: Comparison with Photometric and Spectroscopic Observations}",
-      journal = {\apj},
+        title = {A Self-consistent Cloud Model for Brown Dwarfs and Young Giant Exoplanets: Comparison with Photometric and Spectroscopic Observations},
+      journal = {The Astrophysical Journal},
      keywords = {brown dwarfs, planets and satellites: atmospheres, planets and satellites: gaseous planets, Astrophysics - Earth and Planetary Astrophysics},
          year = 2018,
         month = feb,
@@ -85,7 +85,7 @@ archivePrefix = {arXiv},
 @ARTICLE{samland17,
[…]
