### 2025-03-20 #7288 @warrickball
Hi @lorenaacuna, and thanks for a very high-quality article! I've opened [a pull request](https://github.com/lorenaacuna/GASTLI/pull/15) with some editorial tweaks. If they look good to you, I'll start another test run of our publication pipeline.

One point in the text: I noticed the statement

> Neptune’s measured envelope metallicity is 80x solar

How is metallicity defined here? Roughly the lowest solar metallicity I can think of is *Z*=0.0134, for which 80*Z*=1.072. I thought maybe metallicity is being defined by something other than mass abundance but then saw

> 250x solar (80% metal mass fraction)

later. I tried very briefly to make sense of the references for Neptune but they didn't seem to mention metallicity directly. I presume there's just some difference in how stellar physicists and exoplanetologists use the term "metallicity", which might be worth explaining if a stellar physicist picks up this paper.
