### 2025-03-20 59ffa2d
remove optional AAS metadata
