### 2025-03-21 #7288 @lorenaacuna
> I've opened https://github.com/lorenaacuna/GASTLI/pull/15 with some editorial tweaks. If they look good to you, I'll start another test run of our publication pipeline.

Hi @warrickball. The changes in the manuscript look good to me, so I accepted the pull request, thanks.

> How is metallicity defined here? Roughly the lowest solar metallicity I can think of is Z=0.0134, for which 80Z=1.072. I thought maybe metallicity is being defined by something other than mass abundance but then saw

The metallicity in planet science is typically expressed in x (times) solar units, where the elemental abundance (in moles, not mass) of the metals are multiplied by that factor and then equilibrium chemistry calculations estimate the molecular abundances. The other way to express it is in mass fraction of metals. Z = 0.0134 is a metal mass fraction of 1.34%, which is equivalent to 1 x solar in that notation.
