### 2025-03-21 #7288 @warrickball
Thanks for the explanation. I think the use of a number/molar density (rather than mass density) is the critical difference.
