### 2025-03-21 #7288 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
