### 2025-03-21 #7288 @warrickball
Congratulations @lorenaacuna, your paper has been published in JOSS! 🎉
